### 2025-01-15 #174 @editorialbot
[PRE REVIEW]: Structure-based bioinformatics with BiochemicalAlgorithms.jl
**Submitting author:** @jeleclaire (<a href="http://orcid.org/0000-0002-0587-1636">Jennifer Leclaire</a>)
**Repository:** https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.6
**Managing EiC:** Luca Ferranti

Thanks for submitting your paper to JCON @jeleclaire. **Currently, there isn't a JCON editor assigned** to your paper.
