### 2025-01-17 18edfe7
FF: use the `ForceField` float type for parameter file reading
Fixes: #164
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
