### 2025-01-17 6f109dd
FF: read 1-4 ES scaling factor from parameter file
Fixes: #163
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
