### 2025-01-20 7d02cfe
FF: simplify some code
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
