### 2025-01-20 e20af96
FF: silence HB info messages
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
