### 2025-01-21 71aa274
FF: remove unused data files
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
