+++ data/forcefields/AMBER/converter/amberparser.py
@@ -1,227 +1,0 @@
-import re
-import BALL
-from parametrization import *
-
-class AmberParser:
-	def __init__(self):
-		self.LJLineParser = {
-			'RE' : self.__parseLennardJonesRE,
-			'AC' : self.__parseLennardJonesAC,
-			'SK' : self.__parseLennardJonesSK
-		}
-
-	def parse(self, input):
-		self.params = Parametrization()
-		inFile = open(input)
-
-		inFile.readline()
-
-		self.__parseAtomTypes(inFile)
-		self.__parseHydrophilicAtoms(inFile)
-		self.__parseQuadraticBondStretch(inFile)
-		self.__parseQuadraticAngleBend(inFile)
-		self.__parseTorsions(inFile)
-		self.__parseImproperTorsions(inFile)
-		self.__parseHydrogenBonds(inFile)
-		self.__parseEquivalencyLists(inFile)
-
-		self.__parseLennardJones(inFile)
-
-		return self.params
-
-
-	def __readPartners(self, str, num):
-		""" This method parses strings of the form: A - B - C ...
-		    Where num is the number of parts seperated by '-' and
-	      str is the number of input strings. It returns a pair
-	      consisting of the splitted A, B, C and the remaining string
-		"""
-		parts = str.split("-", num - 1)
-		next = parts[num - 1].strip().split(" ", 1)
-
-		out = [ x.strip() for x in parts[:-1]]
-		out.append(next[0].strip())
-
-		return (out, next[1])
-
-	def __parseAtomTypes(self, input):
-		fields = input.readline().split()
-		while fields != []:
-			atom = self.params.getAtom(fields[0])
-			atom.mass = float(fields[1])
-			fields = input.readline().split()
-
-	def __parseHydrophilicAtoms(self, input):
-		self.params.hydrophilic_atoms += input.readline().split()
-
-	def __parseQuadraticBondStretch(self, input):
-		line = input.readline()
-		while line.strip() != "":
-			partners, str = self.__readPartners(line, 2)
-			fields = str.split()
-			self.params.bond_stretch[tuple(partners)] = tuple(map(float, fields[:2]))
-			line = input.readline()
-
-	def __parseQuadraticAngleBend(self, input):
-		line = input.readline()
-		while line.strip() != "":
-			partners, str = self.__readPartners(line, 3)
-			fields = str.split()
-			self.params.angle_bends[tuple(partners)] = tuple(map(float, fields[:2]))
-			line = input.readline()
-
-	def __parseTorsions(self, input):
-		line = input.readline()
-		while line.strip() != "":
-			partners, str = self.__readPartners(line, 4)
-			partners = tuple([ x == "X" and "*" or x for x in partners ])
-			fields = str.split()
-			tors = self.params.getTorsion(partners)
-			tors.params.append((int  (fields[0]),
-                          float(fields[1]),
-                          float(fields[2]),
-                          int(float(fields[3]))
-                        ))
-			line = input.readline()
-
-	def __parseImproperTorsions(self, input):
-		line = input.readline()
-		while line.strip():
-			partners, str = self.__readPartners(line, 4)
-			partners = tuple([ x == "X" and "*" or x for x in partners ])
-			fields = str.split()
-			tors = self.params.getImproperTorsion(partners)
-			tors.params.append((float(fields[0]),
-                          float(fields[1]),
-                          int(float(fields[2]))
-                        ))
-			line = input.readline()
-
-	def __parseHydrogenBonds(self, input):
-		line = input.readline()
-		while line.strip():
[…]
