+++ data/forcefields/AMBER/converter/parametrization.py
@@ -1,212 +1,0 @@
-class Atom:
-	pass
-
-class Torsion:
-	def __init__(self):
-		self.params = []
-
-class HBond:
-	def __init__(self, p):
-		self.i = p[0]
-		self.j = p[1]
-		self.a = float(p[2])
-		self.b = float(p[3])
-class ResidueParams:
-	def __init__(self):
-		self.atoms = {}
-		self.impropers = []
-
-	def write(self, output):
-		self.__writeImpropers(output)
-		self.__writeCharges(output)
-
-	def getAtom(self, name):
-		if not self.atoms.has_key(name):
-			self.atoms[name] = Atom()
-		return self.atoms[name]
-
-	def __writeImpropers(self, output):
-		output.write("[ResidueImproperTorsions]\n")
-		output.write("key:name\n")
-		output.write(";\n;\n;    res:atom\n;  -------------\n")
-
-		for i in self.impropers:
-			output.write("   " + i + "\n")
-
-		output.write("\n")
-
-	def __writeCharges(self, output):
-		output.write("[ChargesAndTypeNames]\n")
-		output.write("ver:version key:name value:q value:type\n")
-		output.write("@unit_q=e0\n")
-
-		for k, v in self.atoms.iteritems():
-			output.write("   1.0 %s     %3.5f %s\n" % (k, v.charge, v.typename))
-		output.write("\n")
-
-class Parametrization:
-	atoms = {}
-	torsions = {}
-	equivalency_lists = {}
-	improper_torsions = {}
-	bond_stretch = {}
-	angle_bends = {}
-	hydrophilic_atoms = []
-	h_bonds = []
-	lj_potential = {}
-
-	def getAtom(self, name):
-		if not self.atoms.has_key(name):
-			self.atoms[name] = Atom()
-		return self.atoms[name]
-
-	def getTorsion(self, partners):
-		if not self.torsions.has_key(partners):
-			self.torsions[partners] = Torsion()
-		return self.torsions[partners]
-
-	def getImproperTorsion(self, partners):
-		if not self.improper_torsions.has_key(partners):
-			self.improper_torsions[partners] = Torsion()
-		return self.improper_torsions[partners]
-
-	def write(self, outFile):
-		outFile.write("[Options]\n")
-		outFile.write("key:0\n")
-		outFile.write("@SCEE=1.2\n")
-
-		self.__writeAtomTypes(outFile)
-		self.__writeQuadraticBondStretch(outFile)
-		self.__writeTorsions(outFile)
-		self.__writeImproperTorsions(outFile)
-		self.__writeAngleBend(outFile)
-		self.__writeHydrophilicAtoms(outFile)
-		self.__writeHydrogenBonds(outFile)
-		self.__writeLennardJones(outFile)
-
-	def __writeAtomTypes(self, output):
-		output.write("[AtomTypes]\n")
-		output.write("value:ver key:type value:mass value:comment\n")
-		output.write("@unit=g/mol")
-		output.write(";\n;\n;   Rev Type    mass      comment\n")
-		output.write(";   --- ---- ---------- ----------------------------------\n")
-
-		for k, v in self.atoms.iteritems():
-			output.write("    1.0  %-2s  %4.5f   \"\"\n" % (k, v.mass))
-			if self.equivalency_lists.has_key(k):
-				for i in self.equivalency_lists[k]:
-					if not self.atoms.has_key(i):
-						output.write("    1.0  %-2s  %4.5f   \"\"\n" % (i, v.mass))
-
-		output.write("\n")
-
-	def __writeQuadraticBondStretch(self, output):
-		output.write("[QuadraticBondStretch]\n")
[…]
