### 2025-01-21 daa3f5b
FF: reset unassigned LJ interactions on ff update
Fixes: #170
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
