### 2025-01-21 ffb115c
FF: include all warnings in message shown after ff setup
Fixes: #169
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
