+++ src/forcefields/common/forcefield.jl
@@ -25,6 +25,22 @@ abstract type AbstractForceFieldComponent{T<:Real} end
     energy::Dict{String, T}
     unassigned_atoms::Vector{Atom{T}}
     constrained_atoms::Vector{Int}
+    warnings::Vector{String}
+
+    function ForceField{T}(
+        name::String,
+        system::AbstractAtomContainer{T},
+        parameters::AbstractForceFieldParameters,
+        options::Dict{Symbol, Any} = Dict{Symbol, Any}(),
+        atom_type_templates::Dict{String, AtomTypeTemplate{T}} = Dict{String, AtomTypeTemplate{T}}(),
+        components::Vector{AbstractForceFieldComponent{T}} = Vector{AbstractForceFieldComponent{T}}(),
+        energy::Dict{String, T} = Dict{String, T}(),
+        unassigned_atoms::Vector{Atom{T}} = Vector{Atom{T}}(),
+        constrained_atoms::Vector{Int} = Vector{Int}(),
+        warnings::Vector{String} = Vector{String}()
+    ) where T
+        new{T}(name, system, parameters, options, atom_type_templates, components, energy, unassigned_atoms,constrained_atoms, warnings)
+    end
 end
 
 function init_atom_types(params::AbstractForceFieldParameters, ::Type{T}=Float32) where {T <: Real}
@@ -119,11 +135,14 @@ function assign_typenames_and_charges!(ff::ForceField)
                     overwrite_charges=overwrite_charges
                 )
                     # ok, we really don't know the atom type
-                    @warn "assign_typenames_and_charges!(): cannot assign type and/or charge for atom $(get_full_name(atom))"
+                    push!(
+                        ff.warnings,
+                        "assign_typenames_and_charges!(): cannot assign type and/or charge for atom $(get_full_name(atom))"
+                    )
 
                     push!(ff.unassigned_atoms, atom)
                     if length(ff.unassigned_atoms) > ff.options[:max_number_of_unassigned_atoms]
-                        @warn "assigned_typenames_and_charges!(): Too many unassigned atoms"
+                        @error "assigned_typenames_and_charges!(): Too many unassigned atoms"
                         throw(TooManyErrors())
                     end
                 end
@@ -139,28 +158,7 @@ function print_warnings(::AbstractForceFieldComponent) end
 
 function setup!(ff::ForceField)
     map(setup!, ff.components)
-
-    warning_counts = map(count_warnings, ff.components)
-
-    if sum(warning_counts) > 0
-
-        max_length = maximum(length(c.name) for (i,c) in enumerate(ff.components) if warning_counts[i] > 0)
-
-        warning_string = "$(sum(warning_counts)) warnings occurred during setup that were suppressed:\n"
-        warning_string *= "Components:\n"
-        for (i, c) in enumerate(ff.components)
-            if warning_counts[i] > 0
-                warning_string *= Printf.format(
-                    Printf.Format("%-$(max_length)s: %d warnings\n"),
-                        c.name,
-                        warning_counts[i]
-                )
-            end
-        end
-        warning_string *= "Use print_warnings(ff) to display them."
-
-        @warn warning_string
-    end
+    nothing
 end
 
 """
@@ -213,10 +211,50 @@ function compute_forces!(ff::ForceField{T}) where T
     nothing
 end
 
-function print_warnings(ff::ForceField)
-    for component in ff.components
-        print_warnings(component)
+function _check_warnings(ff::ForceField)
+    nwarnings = count_warnings(ff)
+    warning_counts = _warning_counts(ff)
+
+    if nwarnings > 0
+        max_length = maximum(length(name) for (name, cnt) in warning_counts if cnt > 0)
+
[…]
