### 2025-01-22 479c70c
v0.5.0
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
