### 2025-01-31 4e10d8c
System: allow system merging through `Base.append!`
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
