### 2025-01-31 a958350
System: extend `Base.show` for systems, include more information
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
