### 2025-01-31 a60f4dd
System: enable markdown backend for pretty component tables
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
