+++ src/preprocessing/add_hydrogens.jl
@@ -12,7 +12,7 @@ function place_peptide_bond_h_!(frag::Fragment{T}) where T
 
         if isnothing(prev_residue) # should not happen, as we are not the n-terminal
             @warn "Inconsistent chain encountered!"
-        
+
             return false
         end
 
@@ -34,20 +34,15 @@ function place_peptide_bond_h_!(frag::Fragment{T}) where T
 
         # add the new hydrogen
         h_atom = Atom(
-            frag, 
+            frag,
             maximum(atoms(parent_system(frag)).number)+1, # does this make sense?
             Elements.H,
-            "H",
-            "",
-            n_atom.r - (OC * BOND_LENGTH_N_H) / length,
-            zero(Vector3{T}),
-            zero(Vector3{T}),
-            0,
-            zero(T),
-            zero(T),
+            name = "H",
+            atom_type = "",
+            r = n_atom.r - (OC * BOND_LENGTH_N_H) / length,
         )
 
-        Bond(parent_system(frag), n_atom.idx, h_atom.idx, BondOrder.Single, Properties())
+        Bond(parent_system(frag), n_atom.idx, h_atom.idx, BondOrder.Single)
 
         return true
     end
@@ -58,9 +53,9 @@ end
 function _get_connectivity(atom)
     try
        element = atom.element
- 
+
         group = chem_elements[Symbol(element)].group.no
-    
+
         if group < 1 || (group > 2 && group < 13)
             return 0
         end
@@ -110,22 +105,22 @@ function _get_bond_length(element::ElementType, mmff_params::MMFF94Parameters{T}
     if (radius == 0)
         return one(T);
     end
-	
+
     radius_h = mmff_params.radii[Int(Elements.H)]
 
     #  c and n are constants defined in R.Blom and A. Haaland,
     #  J. Molec. Struc, 1985, 128, 21-27.
     # calculate proportionality constant c
     c = T(0.05)
-   
+
     # POWER
 	n = T(1.4)
 
     diff_e = abs(
-        mmff_params.electronegativities[Int(Elements.H)] - 
+        mmff_params.electronegativities[Int(Elements.H)] -
         mmff_params.electronegativities[Int(element)])
 
-    # FORMULA 
+    # FORMULA
 	radius + radius_h - c * diff_e^n
 end
 
@@ -133,20 +128,15 @@ end
 function _add_hydrogen!(atom::Atom{T}, r::Vector3{T}, atom_nr::Int) where {T<:Real}
     # add the new hydrogen
     h_atom = Atom(
-        parent_fragment(atom), 
+        parent_fragment(atom),
         maximum(atoms(parent_system(atom)).number)+1, # does this make sense?
         Elements.H,
-        "H" * ((atom_nr > 1) ? (string(atom_nr) * atom.name) : atom.name),
-        "",
-        r,
-        zero(Vector3{T}),
-        zero(Vector3{T}),
-        0,
-        zero(T),
-        zero(T)
+        name = "H" * ((atom_nr > 1) ? (string(atom_nr) * atom.name) : atom.name),
+        atom_type = "",
+        r = r
     )
 
-    Bond(parent_system(atom), atom.idx, h_atom.idx, BondOrder.Single, Properties())
+    Bond(parent_system(atom), atom.idx, h_atom.idx, BondOrder.Single)
 end
 
 function _get_normal(v::Vector3{T}) where {T<:Real}
@@ -168,48 +158,48 @@ function _handle_atom!(atom::Atom{T}, mmff_params::MMFF94Parameters{T}, is_ring_
     if nbonds(atom) == 1 && first(bonds(atom)).order == BondOrder.Aromatic
         return 0, atom_nr
     end
-    
+
[…]
