### 2025-03-21 3405c2f
minimimalist test
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
