### 2025-03-25 c16258d
Fix `add_hydrogens` (#178)
* closes #177

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* minimimalist test

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* Update test/preprocessing/test_add_hydrogens.jl

Co-authored-by: Thomas Kemmer <thomas@binaryden.de>
Signed-off-by: jeleclaire <leclaire@uni-mainz.de>

* update and accepted requested change

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* Update src/preprocessing/add_hydrogens.jl

Co-authored-by: Thomas Kemmer <thomas@binaryden.de>
Signed-off-by: jeleclaire <leclaire@uni-mainz.de>

---------

Co-authored-by: Tho
[…]
