+++ paper/ref.bib
@@ -194,7 +194,6 @@
   note = {Accessed: 2024-12-16}
 }
 
-
 @misc{BALLTutorial,
   title = {BALL Project Tutorial},
   author = {{BALL Project Contributors}},
@@ -202,7 +201,6 @@
   howpublished = {\url{https://github.com/BALL-Project/ball/blob/master/doc/TUTORIAL/}},
   note = {Accessed: 2024-12-16}
 }
-
 
 @article{Ousterhout1998,
   author = {Ousterhout, John K.},
@@ -212,10 +210,9 @@
   number = {3},
   pages = {23-30},
   year = {1998},
-  publisher = {IEEE Computer Society}
+  publisher = {IEEE Computer Society},
+  doi={10.1109/2.660187}}
 }
-
-
 
 @ARTICLE{BenchmarkTools.jl-2016,
   author =	 {{Chen}, Jiahao and {Revels}, Jarrett},
@@ -246,6 +243,7 @@
   howpublished = {\url{https://github.com/julia-vscode/TestItemRunner.jl}},
   note = {Julia package}
 }
+
 @article{Bouchet-Valat2023,
   title={DataFrames.jl: Flexible and Fast Tabular Data in Julia},
   author={Bouchet-Valat, Milan and Kamiński, Bogumił},
@@ -257,18 +255,16 @@
   doi={10.18637/jss.v107.i04},
   url={https://www.jstatsoft.org/index.php/jss/article/view/v107i04}
 }
-@article{BouchetValat2018,
-  title={Tables.jl: A table interface for everyone},
-  author={Bouchet-Valat, Milan and Quinn, Jacob},
-  journal={Journal of Open Source Software},
-  volume={3},
-  number={25},
-  pages={776},
-  year={2018},
-  doi={10.21105/joss.00776}
+
+@software{BouchetValat2018,
+	title = {{JuliaData}/{Tables}.jl: v1.12.0},
+	url = {https://doi.org/10.5281/zenodo.12753139},
+	publisher = {Zenodo},
+	author = {Quinn, Jacob and Kamiński, Bogumił and Anthoff, David and Bouchet-Valat, Milan and Papp, Tamas K. and Arakaki, Takafumi and Schouten, Rafael and Robinson, Nick and {mathieu17g} and Samuel, Okon and Revels, Jarrett and {ExpandingMan} and Hanson, Eric and Anthony Blaom, PhD and Arslan, Alex and Ling, Jerry and Chen, Jiahao and Day, Josh and Calderón, José Bayoán Santiago and TagBot, Julia and Carlsson, Kristoffer and Hafner, Lilith Orion and Cranmer, Miles and Zwitch, Randy and Johnson, Steven G. and Gillam, Tom and Yang, Yue and {spaette} and {strickek} and Adenbaum, Jacob},
+	month = jul,
+	year = {2024},
+	doi = {10.5281/zenodo.12753139},
 }
-
-
 
 @book{gamma1994design,
 	title={Design Patterns: Elements of Reusable Object-Oriented Software},
@@ -278,7 +274,6 @@
 	address={Reading, Massachusetts},
 	isbn={0-201-63361-2}
 }
-
 
 @software{ProtoSyn,
 	author = {Santos-Martins, Diogo and Sousa, Sérgio F.},
