### 2025-03-25 0ba5537
CI: add Julia 1.12-nightly
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
