+++ .github/workflows/Downgrade.yml
@@ -21,12 +21,15 @@ jobs:
         version:
           - 'lts'
           - '1.11'
+          - '1.12-nightly'
           - 'nightly'
         os:
           - ubuntu-latest
         arch:
           - x64
         include:
+          - version: '1.12-nightly'
+            experimental: true
           - version: 'nightly'
             experimental: true
     steps:
