### 2025-04-04 9787d89
CI: use Julia pre instead of 1.12-nightly
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
