+++ .github/workflows/CI.yml
@@ -31,7 +31,7 @@ jobs:
         version:
           - 'lts'
           - '1.11'
-          - '1.12-nightly'
+          - 'pre'
           - 'nightly'
         os:
           - ubuntu-latest
@@ -40,7 +40,7 @@ jobs:
           # Broken by dependency, cf. https://github.com/BioJulia/IndexableBitVectors.jl/issues/9
           # - x86
         include:
-          - version: '1.12-nightly'
+          - version: 'pre'
             experimental: true
           - version: 'nightly'
             experimental: true
