### 2025-04-04 c7011f8
CI: use Julia pre instead of 1.12-nightly
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
