+++ .github/workflows/Downgrade.yml
@@ -21,14 +21,14 @@ jobs:
         version:
           - 'lts'
           - '1.11'
-          - '1.12-nightly'
+          - 'pre'
           - 'nightly'
         os:
           - ubuntu-latest
         arch:
           - x64
         include:
-          - version: '1.12-nightly'
+          - version: 'pre'
             experimental: true
           - version: 'nightly'
             experimental: true
