### 2025-04-08 6c493ec
finished set_torsion and added missing use cases
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
