+++ test/core/test_fragment.jl
@@ -1418,6 +1418,58 @@ end
     end
 end
 
+@testitem "Fragment/has_torsion" begin
+    sys = load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
+    fdb = FragmentDB()
+    build_bonds!(sys, fdb)
+    r1 = residues(sys)[1]
+    r2 = residues(sys)[2]
+
+    @test has_torsion_psi(r1) == true
+    @test has_torsion_psi(r2) == true
+
+    @test has_torsion_phi(r1) == true
+    @test has_torsion_phi(r2) == true
+    @test has_torsion_omega(r1) == true
+    @test has_torsion_omega(r2) == true
+
+    sys2 = deepcopy(sys)
+    @test nfragments(sys2) == 2
+    @test nresidues(sys2) == 2
+    delete!(residues(sys2)[2])
+    @test nresidues(sys2) == 1
+
+    r1 = residues(sys2)[1]
+    @test has_torsion_psi(r1) == false
+    @test has_torsion_phi(r1) == false
+    @test has_torsion_omega(r1) == false
+
+    r1 = residues(sys)[1]
+    r2 = residues(sys)[2]
+
+    @test !has_flag(r1, :N_TERMINAL)
+    @test !has_flag(r1, :C_TERMINAL)
+
+    @test !has_flag(r2, :N_TERMINAL)
+    @test !has_flag(r2, :C_TERMINAL)
+
+    set_flag!(r1, :N_TERMINAL)
+    @test has_flag(r1, :N_TERMINAL)
+
+    set_flag!(r2, :C_TERMINAL)
+    @test has_flag(r2, :C_TERMINAL)
+
+    @test has_torsion_psi(r1) == true
+    @test has_torsion_psi(r2) == false
+
+    @test has_torsion_phi(r1) == false
+    @test has_torsion_phi(r2) == true
+    @test has_torsion_omega(r1) == false
+    @test has_torsion_omega(r2) == true
+
+end
+
+
 @testitem "Fragment/calculate_torsion_angle" begin
     sys = System()
     a = Atom(sys, 1, Elements.C, r = Vector3{Float32}(0.0f0, 1.0f0, 0.0f0))
@@ -1440,4 +1492,35 @@ end
 
     a.r = b.r
     @test_logs (:error, "Atoms can't have the same position. No Angle to calculate here.") calculate_bond_angle(a,b,c)
+end
+
+@testitem "Fragment/set_torsion_angle" begin
+    sys = System()
+    a = Atom(sys, 1, Elements.C, r = Vector3{Float32}(0.0, 1.0, 0.0))
+    b = Atom(sys, 2, Elements.C, r = Vector3{Float32}(0.00, 0.0, 0.0))
+    c = Atom(sys, 3, Elements.N, r = Vector3{Float32}(1.0, 0.0, 0.0))
+    d = Atom(sys, 4, Elements.C, r = Vector3{Float32}(1.0, -1.0, 0.0))
+
+    Bond(a, b, BondOrder.Single)
+    Bond(b, c, BondOrder.Single)
+    Bond(c, d, BondOrder.Single)
+
+    @test set_torsion_angle!(a, b, c, d, deg2rad(90.0))
+    @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 90.000, atol=10e-3)
+
+    @test set_torsion_angle!(a, b, c, d, deg2rad(0.0))
+    @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 0.0, atol=10e-3)
+
+    set_torsion_angle!(a, b, c, d, deg2rad(149.99))
[…]
