### 2025-04-08 ac1b7a4
finished set_torsion and added missing use cases
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
