### 2025-04-09 42c4255
restructuring of tutorials
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
