### 2025-04-09 b5441eb
restructuring of tutorials
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
