+++ src/fileformats/pdb/pdb_writer.jl
@@ -1,0 +1,517 @@
+using Printf
+using Dates
+using DataFrames
+
+using BiochemicalAlgorithms
+
+function extract_structure(ac::AtomContainer)
+    # TODO: handle crystal information, like in the original BALL PDB writer
+
+    # Walk over all atoms and collect all chains, bonds, etc...
+
+
+end
+
+# this really should be in the standard library
+unval(::Val{x}) where x = x
+
+function fix_nucleotide_name(name::String)
+    replace(name, 
+            r"^DA$" => "A",
+            r"^DC$" => "C",
+            r"^DG$" => "G",
+            r"^DT$" => "T",
+            r"^DI$" => "I",
+            r"^DU$" => "U")
+end
+
+function residue_details(res::Fragment{T}) where {T <:Real}
+    res.name, isnothing(parent_chain(res)) ? "" : parent_chain(res).name, res.number, get_property(res, :insertion_code, " ")
+end
+
+function atom_details(atom::Atom{T}) where {T <:Real}
+    res = parent_fragment(atom)
+
+    if isnothing(res)
+        "UNK", "", 0, ""
+    else
+        residue_details(res)
+    end
+end
+
+function extract_records(record_type::Val, pdb_info::PDBInfo)
+    records = Iterators.filter(r -> Val(r.type) == record_type, pdb_info.records)
+end
+
+function break_into_continuation_lines(line::String, c::Int = 1)
+    if length(line) > 70 # we need to leave room for the eol, so we skip one character at the end
+        [(c, line[1:69]), break_into_continuation_lines(String(line[71:end]), c+1)...]
+    else
+        [(c, line)]
+    end
+end
+
+function write_record(io::IO, pdb_info::PDBInfo, tag_name::String, args...)
+    type = RECORD_MAP[tag_name]
+    
+    # CONECT-records are handled differently
+    if type.record_type === RECORD_TYPE__REMARK
+        pdb_info.writer_stats.remark_records += 1
+    elseif type.record_type === RECORD_TYPE__HET
+        pdb_info.writer_stats.het_records += 1
+    elseif type.record_type === RECORD_TYPE__HELIX
+        pdb_info.writer_stats.helix_records += 1
+    elseif type.record_type === RECORD_TYPE__SHEET
+        pdb_info.writer_stats.sheet_records += 1
+    elseif type.record_type === RECORD_TYPE__TURN
+        pdb_info.writer_stats.turn_records += 1
+    elseif type.record_type === RECORD_TYPE__SITE
+        pdb_info.writer_stats.site_records += 1
+    elseif type.record_type in [
+            RECORD_TYPE__ORIGX1, RECORD_TYPE__ORIGX2, RECORD_TYPE__ORIGX3,
+            RECORD_TYPE__SCALE1, RECORD_TYPE__SCALE2, RECORD_TYPE__SCALE3,
+            RECORD_TYPE__MTRIX1, RECORD_TYPE__MTRIX2, RECORD_TYPE__MTRIX3    
+        ]
+        pdb_info.writer_stats.coordinate_transformation_records += 1
+    elseif type.record_type in [RECORD_TYPE__ATOM, RECORD_TYPE__HETATM]
+        pdb_info.writer_stats.atomic_coordinate_records += 1
+    elseif type.record_type == RECORD_TYPE__TER
+        pdb_info.writer_stats.ter_records += 1
+    else type.record_type == RECORD_TYPE__SEQRES
+        pdb_info.writer_stats.seqres_records += 1
+    end
+
+    Printf.format(io, Printf.Format(type.tag * type.format_string * "\n"), args...)
+end
+
[…]
