+++ Project.toml
@@ -11,6 +11,7 @@ CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 CellListMap = "69e1c6dd-3888-40e6-b3c8-31ac5f578864"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 EnumX = "4e289a0a-7415-4d19-859d-a7e5c4648b56"
 GraphDataFrameBridge = "3c71623a-a715-5176-9801-629b201a4880"
@@ -42,6 +43,7 @@ CSV = "0.10"
 CellListMap = "0.8, 0.9"
 DataFrames = "1"
 DataStructures = "0.18"
+Dates = "1.11.0"
 DocStringExtensions = "0.9"
 EnumX = "1"
 GraphDataFrameBridge = "0.3"
