+++ src/fileformats/pdb/pdb_defs.jl
@@ -94,6 +94,9 @@ end
     number::Int
     first::UniqueResidueID
     second::UniqueResidueID
+    symmetry_operator_0::Int
+    symmetry_operator_1::Int
+    bond_length::Real
 end
 
 @auto_hash_equals struct HelixRecord
@@ -110,7 +113,7 @@ end
     name::String
     initial_residue::UniqueResidueID
     terminal_residue::UniqueResidueID
-    sense_of_strand::Bool
+    sense_of_strand::Int
 end
 
 @auto_hash_equals struct TurnRecord
@@ -119,6 +122,24 @@ end
     initial_residue::UniqueResidueID
     terminal_residue::UniqueResidueID
     comment::String
+end
+
+@auto_hash_equals mutable struct PDBWriterStats
+    remark_records::Int
+    het_records::Int
+    helix_records::Int
+    sheet_records::Int
+    turn_records::Int
+    site_records::Int
+    coordinate_transformation_records::Int
+    atomic_coordinate_records::Int
+    ter_records::Int
+    conect_records::Int
+    seqres_records::Int
+
+    function PDBWriterStats()
+        new(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
+    end
 end
 
 @auto_hash_equals mutable struct PDBInfo{T}
@@ -142,10 +163,12 @@ end
 
     alternate_location_identifier::String
 
+    writer_stats::PDBWriterStats
+
     function PDBInfo{T}(selected_model=-1) where {T}
         new("", "", "", "", Deque{PDBRecord}(), Deque{SSBondRecord}(), 
             Deque{Union{HelixRecord, SheetRecord, TurnRecord}}(), selected_model, 
-            1, nothing, nothing, "A")
+            1, nothing, nothing, "A", PDBWriterStats())
     end
 end
 
@@ -154,10 +177,10 @@ const FORMAT_UNKNOWN          = ""
 const FORMAT_ANISOU           = "%5ld %-4.4s%c%3.3s %c%4ld%c %7ld%7ld%7ld%7ld%7ld%7ld  %4.4s%2.2s%2.2s"
 const FORMAT_ATOM             = "%5ld %-4.4s%c%3.3s %c%4ld%c   %8.3f%8.3f%8.3f%6.2f%6.2f      %4.4s%2.2s%2.2s"
 const FORMAT_ATOM_PARTIAL_CRG = "%5ld %-4.4s%c%3.3s %c%4ld%c   %8.3f%8.3f%8.3f%6.2f%6.2f      %4.4s%4.4s"
-const FORMAT_AUTHOR           = "  %2ld%-60.60s"
-const FORMAT_CAVEAT           = "  %2ld %4.4s    %51.51s"
+const FORMAT_AUTHOR           = "  %2.2s%-60.60s"
+const FORMAT_CAVEAT           = "  %2.2s %4.4s    %-51.51s"
 const FORMAT_CISPEP           = " %3ld %3.3s %c %4ld%c   %3.3s %c %4ld%c       %3ld       %6f"
-const FORMAT_COMPND           = "  %2ld%-60.60s"
+const FORMAT_COMPND           = "  %2.2s%-60.60s"
 const FORMAT_CON06            = "%5ld%5ld%5ld%5ld%5ld"
 const FORMAT_CON061           = "%5ld"
[…]
