+++ .github/workflows/Downgrade.yml
@@ -41,7 +41,7 @@ jobs:
       - uses: julia-actions/cache@v2
       - uses: julia-actions/julia-downgrade-compat@v1
         with:
-          skip: LinearAlgebra,Printf,Statistics,Test
+          skip: Dates,LinearAlgebra,Printf,Statistics,Test
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
         with:
