+++ test/core/test_system.jl
@@ -6,23 +6,16 @@
         reconstruct_fragments!(testsys, fdb)
         build_bonds!(testsys, fdb)
 
-        # TODO: currently, the PDB parser does not yet create SecondaryStructures
-        #       for testing purposes, we just add one to each chain and add all fragments of the chain into it
-        for c in chains(testsys)
-            ss = SecondaryStructure(c, 1, SecondaryStructureElement.Coil; name = c.name)
-            fragments(c).secondary_structure_idx .= Ref(ss.idx)
-        end
-
-        @test natoms(testsys) == 3790
+        @test natoms(testsys) == 3778
         @test nbonds(testsys) == 3839
         @test nmolecules(testsys) == 1
-        @test nchains(testsys) == 3
-        @test nsecondary_structures(testsys) == 3
+        @test nchains(testsys) == 6
+        @test nsecondary_structures(testsys) == 30
         ct = chains(testsys)
-        @test nfragments.(ct) == [192, 17, 13]
-        @test nfragments.(ct; variant = FragmentVariant.None) == [12, 5, 1]
-        @test nnucleotides.(ct) == [0, 12, 12]
-        @test nresidues.(ct) == [180, 0, 0]
+        @test nfragments.(ct) == [12, 12, 180, 5, 1, 12]
+        @test nfragments.(ct; variant = FragmentVariant.None) == [0, 0, 0, 5, 1, 12]
+        @test nnucleotides.(ct) == [0, 0, 0, 0, 0, 0]
+        @test nresidues.(ct) == [12, 12, 180, 0, 0, 0]
 
         # empty!
         sys = deepcopy(testsys)
@@ -104,9 +97,9 @@
         aidx = Set(at.idx)
         bidx = Set(bonds(at).idx)
         @test delete!(at) === at
-        @test natoms(sys) == 3739
+        @test natoms(sys) == 3727
         @test length(at) == 0
-        @test nbonds(sys) == 3712
+        @test nbonds(sys) == 3721
         @test length(aidx ∩ Set(atoms(sys).idx)) == 0
         @test length(bidx ∩ Set(bonds(sys).idx)) == 0
 
@@ -119,9 +112,9 @@
         bidx = Set(bonds(atom).idx)
         @test_throws KeyError delete!(at, -1)
         @test delete!(at, aidx) === at
-        @test natoms(sys) == 3738
-        @test length(at) == 3738
-        @test nbonds(sys) == 3708
+        @test natoms(sys) == 3726
+        @test length(at) == 3726
+        @test nbonds(sys) == 3719
         @test aidx ∉ atoms(sys).idx
         @test length(bidx ∩ Set(bonds(sys).idx)) == 0
 
@@ -129,8 +122,8 @@
         @test_throws KeyError first(stale_col)
         @test revalidate_indices!(stale_table) === stale_table
         @test revalidate_indices!(stale_col) === stale_col
-        @test length(stale_table) == 3738
-        @test length(stale_col) == 3738
+        @test length(stale_table) == 3726
+        @test length(stale_col) == 3726
 
         @test delete!(at) === at
         @test natoms(sys) == 0
@@ -227,7 +220,7 @@
         @test_throws KeyError delete!(mt, -1)
         @test delete!(mt, first(mt.idx); keep_atoms = true) === mt
         @test length(mt) == 0
-        @test natoms(sys) == 3790
+        @test natoms(sys) == 3778
         @test all(isnothing, atoms(sys).molecule_idx)
         @test nbonds(sys) == 3839
         @test nmolecules(sys) == 0
@@ -239,7 +232,7 @@
         mt = molecules(sys)
         @test delete!(mt; keep_atoms = true) === mt
         @test length(mt) == 0
-        @test natoms(sys) == 3790
+        @test natoms(sys) == 3778
[…]
