### 2025-04-11 1b9f3ef
additional tests and missing docs added
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
