### 2025-04-11 8fbd1a0
Apply suggestions from code review
Co-authored-by: Thomas Kemmer <thomas@binaryden.de>
Signed-off-by: jeleclaire <leclaire@uni-mainz.de>
