### 2025-04-14 2395ad9
added missing documentation, test cases for return types and fix of errors
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
