+++ test/core/test_fragment.jl
@@ -1419,55 +1419,56 @@ end
 end
 
 @testitem "Fragment/has_torsion" begin
+    for T in [Float32, Float64]
+        sys = load_pdb(ball_data_path("../test/data/AlaAla.pdb"), T)
+        fdb = FragmentDB{T}()
 
-    sys = load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
-    fdb = FragmentDB()
-    build_bonds!(sys, fdb)
-    r1 = residues(sys)[1]
-    r2 = residues(sys)[2]
+        build_bonds!(sys, fdb)
+        r1 = residues(sys)[1]
+        r2 = residues(sys)[2]
 
-    @test has_torsion_psi(r1) == true
-    @test has_torsion_psi(r2) == true
+        @test has_torsion_psi(r1) == true
+        @test has_torsion_psi(r2) == true
 
-    @test has_torsion_phi(r1) == true
-    @test has_torsion_phi(r2) == true
-    @test has_torsion_omega(r1) == true
-    @test has_torsion_omega(r2) == true
+        @test has_torsion_phi(r1) == true
+        @test has_torsion_phi(r2) == true
+        @test has_torsion_omega(r1) == true
+        @test has_torsion_omega(r2) == true
 
-    sys2 = deepcopy(sys)
-    @test nfragments(sys2) == 2
-    @test nresidues(sys2) == 2
-    delete!(residues(sys2)[2])
-    @test nresidues(sys2) == 1
+        sys2 = deepcopy(sys)
+        @test nfragments(sys2) == 2
+        @test nresidues(sys2) == 2
+        delete!(residues(sys2)[2])
+        @test nresidues(sys2) == 1
 
-    r1 = residues(sys2)[1]
-    @test has_torsion_psi(r1) == false
-    @test has_torsion_phi(r1) == false
-    @test has_torsion_omega(r1) == false
+        r1 = residues(sys2)[1]
+        @test has_torsion_psi(r1) == false
+        @test has_torsion_phi(r1) == false
+        @test has_torsion_omega(r1) == false
 
-    r1 = residues(sys)[1]
-    r2 = residues(sys)[2]
+        r1 = residues(sys)[1]
+        r2 = residues(sys)[2]
 
-    @test !has_flag(r1, :N_TERMINAL)
-    @test !has_flag(r1, :C_TERMINAL)
+        @test !has_flag(r1, :N_TERMINAL)
+        @test !has_flag(r1, :C_TERMINAL)
 
-    @test !has_flag(r2, :N_TERMINAL)
-    @test !has_flag(r2, :C_TERMINAL)
+        @test !has_flag(r2, :N_TERMINAL)
+        @test !has_flag(r2, :C_TERMINAL)
 
-    set_flag!(r1, :N_TERMINAL)
-    @test has_flag(r1, :N_TERMINAL)
+        set_flag!(r1, :N_TERMINAL)
+        @test has_flag(r1, :N_TERMINAL)
 
-    set_flag!(r2, :C_TERMINAL)
-    @test has_flag(r2, :C_TERMINAL)
+        set_flag!(r2, :C_TERMINAL)
+        @test has_flag(r2, :C_TERMINAL)
 
-    @test has_torsion_psi(r1) == true
-    @test has_torsion_psi(r2) == false
+        @test has_torsion_psi(r1) == true
+        @test has_torsion_psi(r2) == false
 
-    @test has_torsion_phi(r1) == false
-    @test has_torsion_phi(r2) == true
-    @test has_torsion_omega(r1) == false
[…]
