### 2025-04-14 6c07a83
removed redundant test, modified docs
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
