+++ tutorials/02_CoreComponents.qmd
@@ -23,7 +23,7 @@ load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
 ```
 
 
-## Atoms:
+## Atom
 
 We can access the atoms like this:
 ```{julia}
@@ -65,71 +65,13 @@ n2 = atom_by_idx(s, 17)
 for a in sys_atms
     println(a.idx, " ", a.number)
 end
-a = Atom(s, number = 7, element = Elements.N;)
 ```
 
-Next, we will have a look at the fragments.
+Next, we will have a look at `Bond`.
 
-## Fragments:
+## Bond
 
-Fragments represent molecule fragments. In context of proteins or nucleic acids, the fragment can be used to represent residues or nucleotides. See the documentation for the definition of the ```FragmentVariantType``` (https://hildebrandtlab.github.io/BiochemicalAlgorithms.jl/stable/public/system/#BiochemicalAlgorithms.FragmentVariant)
-
-```{julia}
-nfragments(s)
-
-for frag in fragments(s)
-    println(frag.name, " ", frag.idx)
-end
-n2 = atom_by_idx(s, 17)
-
-println("The N atom belongs to residue: ", n2.fragment_idx)
-
-```
-
-We can get the fragment as well directly:
-```{julia}
-fragment = parent_fragment(n2)
-println(fragment)
-res2 = fragment_by_idx(s, 3)
-```
-
-Just like atom indices, fragment indices not necessarily follow a certain order but are unique inside the same system object.
-
-
-```{julia}
-res1 = fragments(s)[1]
-res1.variant # 1 = unknown, 2 = residue, 3 = nucleotide
-```
-
-Alternatively, it is also possible to check if this is a nucleotide:
-
-```{julia}
-println(is_nucleotide(res1))
-
-println("Number of fragments before push: ", nfragments(s))
-push!(chain_by_idx(s, 2), res1)
-println("Number of fragments after push: ", nfragments(s))
-delete!(res1)
-println("Number of fragments after deletion: ", nfragments(s))
-
-for frag in fragments(s)
-    println(frag.name, " ", frag.idx)
-end
-s
-```
-
-```{julia}
-s = load_pdb(ball_data_path("../test/data/AlaAla.pdb")) # restore original data
-res1 = fragments(s)[1]
-println("is residue ", isresidue(res1))
-```
-
-There are more functionalities related to fragments. Have a look at the documentation for ```Fragment```(https://hildebrandtlab.github.io/BiochemicalAlgorithms.jl/stable/public/system/#Fragments-(all-variants)) for more details.
-
-
-## Bonds
-
-Next, we will have a look at bonds. BiochemicalAlgorithms.jl supports different BondTypes:
+BiochemicalAlgorithms.jl supports different BondTypes:
 ```{julia}
 BiochemicalAlgorithms.BondOrderType
 ```
@@ -153,7 +95,7 @@ Pkg.add("BiochemicalVisualization")
 ```{julia}
 using BiochemicalVisualization
 
-ball_and_stick(s)
+# ball_and_stick(s)  uncomment for visualization
 ```
 
 Now, let's create some bonds:
@@ -163,7 +105,7 @@ fdb = FragmentDB() # default FragmentDB
 normalize_names!(s, fdb) # in case our input PDB file uses a strange naming standard
 reconstruct_fragments!(s, fdb) # in case our input file misses some atoms
 build_bonds!(s, fdb) # create the bonds
-#ball_and_stick(s)
+# ball_and_stick(s) uncomment for visualization
 ```
 
 And finally, we do have our bonds. This example demonstrate the importance of visualizing for the understanding of structural data.
@@ -229,16 +171,107 @@ println("The bond length is ", bond_length(bond))
 
[…]
