### 2025-04-14 bc0a230
removed outdated md files
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
