### 2025-04-14 aa7a3d4
removed outdated md files
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
