### 2025-04-14 826c555
removed outdated md files
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
