+++ docs/src/tutorials/handle_molecules.md
@@ -1,201 +1,0 @@
-# All on Handling molecules
-
-
-## How can I create a simple molecule?
-
-``` julia
-# create a system first
-sys = System{Float32}() # this system will be of single precision, i.e., atom positions, velocities...
-
-# as well as a molecule
-h2o = Molecule(sys)
-
-# create system atoms
-o1 = Atom(h2o, 1, Elements.O)
-h1 = Atom(h2o, 2, Elements.H)
-h2 = Atom(h2o, 3, Elements.H)
-
-# set positions of the atoms
-# o1.r = [0, 0, 0]  <-- this is the default value!
-h1.r = [1, 0, 0]
-h2.r = [cos(deg2rad(105)), sin(deg2rad(105)), 0]
-
-# add bonds
-Bond(h2o, o1.idx, h1.idx, BondOrder.Single)
-Bond(h2o, o1.idx, h2.idx, BondOrder.Single)
-
-println("Number of atoms: ", natoms(h2o))
-println("Number of bonds: ", nbonds(h2o))
-```
-
-    Number of atoms: 3
-    Number of bonds: 2
-
-## How can I determine the element of an atom (C, N, …)?
-
-``` julia
-sys = load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
-for atom in atoms(sys)
-    println("Atom no.: $(atom.number), element: $(atom.element)")
-end
-```
-
-    Atom no.: 1, element: N
-    Atom no.: 2, element: C
-    Atom no.: 3, element: C
-    Atom no.: 4, element: O
-    Atom no.: 5, element: H
-    Atom no.: 6, element: H
-    Atom no.: 7, element: H
-    Atom no.: 8, element: H
-    Atom no.: 9, element: H
-    Atom no.: 10, element: H
-    Atom no.: 11, element: C
-    Atom no.: 12, element: H
-    Atom no.: 13, element: N
-    Atom no.: 14, element: C
-    Atom no.: 15, element: C
-    Atom no.: 16, element: H
-    Atom no.: 17, element: H
-    Atom no.: 18, element: H
-    Atom no.: 19, element: H
-    Atom no.: 20, element: C
-    Atom no.: 21, element: H
-    Atom no.: 22, element: O
-    Atom no.: 23, element: O
-
-You can also filter for specific elements:
-
-``` julia
-c_atoms = filter(atom -> atom.element == Elements.C, atoms(sys))
-for atom in c_atoms
-    println("Atom no.: $(atom.number), element: $(atom.element)")
-end
-```
-
-    Atom no.: 2, element: C
-    Atom no.: 3, element: C
-    Atom no.: 11, element: C
-    Atom no.: 14, element: C
-    Atom no.: 15, element: C
-    Atom no.: 20, element: C
-
-## How can I identify backbone atoms?
-
-``` julia
-sys = load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
-for atom in atoms(sys)
-    print(atom.element)
-    if atom.name in ["C", "O", "N", "HA"]
-        print(" <-- This is a backbone atom!")
-    end
-    println()
-end
-```
-
-    N <-- This is a backbone atom!
-    C
-    C <-- This is a backbone atom!
-    O <-- This is a backbone atom!
-    H
-    H
-    H
-    H
-    H
-    H
-    C
-    H <-- This is a backbone atom!
-    N <-- This is a backbone atom!
-    C
-    C <-- This is a backbone atom!
-    H
-    H
-    H
-    H
-    C
-    H <-- This is a backbone atom!
-    O <-- This is a backbone atom!
-    O
-
-## How can I get the one-letter code out of a pdb file?
-
-``` julia
[…]
