### 2025-04-15 b8ed6df
removed refs to non-existing docs
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
