### 2025-04-15 e48eebe
removed refs to non-existing docs
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
