+++ src/fileformats/pdb/pdb_defs.jl
@@ -6,6 +6,8 @@ using BiochemicalAlgorithms:
     Chain,
     Fragment
 
+export PDBInfo
+
 @enum(PDBRecordType,
     RECORD_TYPE__UNKNOWN = 0,
     RECORD_TYPE__ANISOU,
@@ -174,12 +176,12 @@ end
 
 
 const FORMAT_UNKNOWN          = ""
-const FORMAT_ANISOU           = "%5ld %-4.4s%c%3.3s %c%4ld%c %7ld%7ld%7ld%7ld%7ld%7ld  %4.4s%2.2s%2.2s"
-const FORMAT_ATOM             = "%5ld %-4.4s%c%3.3s %c%4ld%c   %8.3f%8.3f%8.3f%6.2f%6.2f      %4.4s%2.2s%2.2s"
-const FORMAT_ATOM_PARTIAL_CRG = "%5ld %-4.4s%c%3.3s %c%4ld%c   %8.3f%8.3f%8.3f%6.2f%6.2f      %4.4s%4.4s"
+const FORMAT_ANISOU           = "%5ld %-4.4s%1.1s%3.3s %1.1s%4ld%1.1s %7ld%7ld%7ld%7ld%7ld%7ld  %4.4s%2.2s%2.2s"
+const FORMAT_ATOM             = "%5ld %-4.4s%1.1s%3.3s %1.1s%4ld%1.1s   %8.3f%8.3f%8.3f%6.2f%6.2f      %4.4s%2.2s%2.2s"
+const FORMAT_ATOM_PARTIAL_CRG = "%5ld %-4.4s%1.1s%3.3s %1.1s%4ld%1.1s   %8.3f%8.3f%8.3f%6.2f%6.2f      %4.4s%4.4s"
 const FORMAT_AUTHOR           = "  %2.2s%-60.60s"
 const FORMAT_CAVEAT           = "  %2.2s %4.4s    %-51.51s"
-const FORMAT_CISPEP           = " %3ld %3.3s %c %4ld%c   %3.3s %c %4ld%c       %3ld       %6f"
+const FORMAT_CISPEP           = " %3ld %3.3s %1.1s %4ld%1.1s   %3.3s %1.1s %4ld%1.1s       %3ld       %6f"
 const FORMAT_COMPND           = "  %2.2s%-60.60s"
 const FORMAT_CON06            = "%5ld%5ld%5ld%5ld%5ld"
 const FORMAT_CON061           = "%5ld"
@@ -188,25 +190,25 @@ const FORMAT_CON063           = "%5ld%5ld%5ld"
 const FORMAT_CON064           = "%5ld%5ld%5ld%5ld"
 const FORMAT_CONECT           = "%5ld%5ld%5ld%5ld%5ld%5ld%5ld%5ld%5ld%5ld%5ld"
 const FORMAT_CRYST1           = "%9.3f%9.3f%9.3f%7.2f%7.2f%7.2f %-11.11s%4ld"
-const FORMAT_DBREF            = " %4.4s %c %4ld%c %4ld%c %6.6s %8.8s %12.12s %5ld%c %5ld%c"
+const FORMAT_DBREF            = " %4.4s %1.1s %4ld%1.1s %4ld%1.1s %6.6s %8.8s %12.12s %5ld%1.1s %5ld%1.1s"
 const FORMAT_END              = ""
 const FORMAT_ENDMDL           = ""
 const FORMAT_EXPDTA           = "  %2.2s%-60.60s"
-const FORMAT_FORMUL           = "  %2.2s  %3.3s %2ld%c%51.51s"
+const FORMAT_FORMUL           = "  %2.2s  %3.3s %2ld%1.1s%51.51s"
[…]
