### 2025-04-15 eee5fae
Add compat header for Scanf
Signed-off-by: Andreas Hildebrandt <andreas.hildebrandt@uni-mainz.de>
