### 2025-04-15 0d5a236
Add compat header for Scanf
Signed-off-by: Andreas Hildebrandt <andreas.hildebrandt@uni-mainz.de>
