+++ Project.toml
@@ -61,6 +61,7 @@ PrettyTables = "2.2.7"
 Printf = "1.10"
 Quaternions = "0.7.6"
 Rotations = "1.3.4"
+Scanf = "0.5.4"
 StaticArrays = "1.5"
 Statistics = "1.10"
 StructTypes = "1.5"
