### 2025-04-16 ea9606c
Fix previous / next in fragment
Signed-off-by: Andreas Hildebrandt <andreas.hildebrandt@uni-mainz.de>
