### 2025-04-16 4824e09
Fix tests for secondary structure prediction
Signed-off-by: Andreas Hildebrandt <andreas.hildebrandt@uni-mainz.de>
