+++ test/preprocessing/test_predict_secondary_structure.jl
@@ -1,68 +1,30 @@
 
 
-@testitem "predict secondary structure for BPTI" begin
+@testitem "predict secondary structure" begin
 
-    default_fdb = FragmentDB()
-    fdb = FragmentDB(ball_data_path("fragments/Fragments.db.json"))
-    sys = load_pdb(ball_data_path("../test/data/PDBFile_test2.pdb"))
-   
-    normalize_names!(sys, fdb)
-    reconstruct_fragments!(sys, fdb)
-    build_bonds!(sys, fdb)
-    predict_hbonds!(sys, :KABSCH_SANDER)
-    summary = predict_secondary_structure!(sys)
-    
-    #@test natoms(sys) == 892 ## vorher 892 TODO checken ob das an Cysteinen liegt
+    function test_secondary_structure(ac::AbstractAtomContainer{T}) where {T}
+        fdb = FragmentDB()
+
+        normalize_names!(ac, fdb)
+        reconstruct_fragments!(ac, fdb)
+        build_bonds!(ac, fdb)
+        predict_hbonds!(ac, :KABSCH_SANDER)
+        predict_secondary_structure!(ac)
+    end
+
+    sys_pdb_test2 = load_pdb(ball_data_path("../test/data/PDBFile_test2.pdb"))
+    @test test_secondary_structure(sys_pdb_test2) == "--GGGG-----------EEEEEEETTTTEEEEEEE---------B--HHHHHHHH---"
 
     # the reference was generated by the C++ BALL version
-    @test String(summary) == "--GGGG-----------EEEEEEETTTTEEEEEEE---------B--HHHHHHHH---"
-    @test length(secondary_structures(sys)) == 9
+    @test length(secondary_structures(sys_pdb_test2)) == 9
     
     # the following was additionally 'compared to' in the C++ version 
 	#String PDB_summary = "CCGGGGSCCCCCSCCCCEEEEEEETTTTEEEEEEECSSSCCSSCBSSHHHHHHHHSCC";
-    
-
-
-    function predict_secondary_structure(sys)
-        default_fdb = FragmentDB()
-        fdb = FragmentDB(ball_data_path("fragments/Fragments.db.json"))
-
-        filter_fn = a -> is_amino_acid(parent_fragment(atom_by_idx(sys, a.idx)))
-
-        filtered_sys = filter_atoms(
-            filter_fn,
-            sys
-        )
-        
-        normalize_names!(filtered_sys, fdb)
-        reconstruct_fragments!(filtered_sys, fdb)
-        build_bonds!(filtered_sys, fdb)
-        predict_hbonds!(filtered_sys, :KABSCH_SANDER)
-        summary = predict_secondary_structure!(filtered_sys)
-
-        return summary
-    end
 
     ### 5pti.pdb
     sys_5pti = load_pdb(ball_data_path("../test/data/5PTI.pdb"))
-    @test String(predict_secondary_structure(sys_5pti)) == 
-          "--GGGG-----------EEEEEEETTTTEEEEEEE---------B--HHHHHHHH---"
-    
-    #### 3fgu.pdb
-    sys_3fgu = load_pdb(ball_data_path("../test/data/3fgu.pdb"))
-    @test String(predict_secondary_structure(sys_3fgu)) ==
[…]
