+++ src/BiochemicalAlgorithms.jl
@@ -105,6 +105,8 @@ include("preprocessing/predict_secondary_structure.jl")
 # optimization
 include("optimization/optimize_structure.jl")
 
+include("peptide_builder.jl")
+
 export
     ball_data_path
 
