### 2025-04-22 dd3c745
CompatHelper: bump compat for MolecularGraph to 0.18, (keep existing compat)
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
