### 2025-04-23 804ab55
CI: exclude CompatHelper from DCO checks
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
