### 2025-04-29 4fad212
Apply suggestions from code review
Co-authored-by: Thomas Kemmer <thomas@binaryden.de>
Signed-off-by: jeleclaire <leclaire@uni-mainz.de>
