### 2025-04-29 f10adf2
Add is_previous and is_next to fragment
Signed-off-by: Andreas Hildebrandt <andreas.hildebrandt@uni-mainz.de>
