### 2025-04-30 457fb99
added functions for calculation of torsion angles given the fragment
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
