+++ src/core/fragment.jl
@@ -1,6 +1,5 @@
 export
     calculate_bond_angle,
-    calculate_torsion_angle,
     Fragment,
     FragmentTable,
     Nucleotide,
@@ -10,6 +9,9 @@ export
     get_full_name,
     get_previous,
     get_next,
+    get_torsion_omega,
+    get_torsion_phi,
+    get_torsion_psi,
     has_torsion_omega,
     has_torsion_phi,
     has_torsion_psi,
@@ -676,13 +678,8 @@ end
 """
     has_torsion_psi(frag::Fragment) -> Bool
 
-Checks if the given `Fragment` object has a torsion angle psi defined.
+Checks if the given `frag::Fragment` object has a torsion angle psi defined, `false` otherwise.
 See [Wikipedia](https://en.wikipedia.org/wiki/File:Protein_backbone_PhiPsiOmega_drawing.svg) for more information.
-# Arguments
-- `frag::Fragment`: The fragment object to be checked.
-
-# Returns
-- `Bool`: `true` if the fragment has a torsion angle psi, `false` otherwise.
 """
 @inline function has_torsion_psi(frag::Fragment)
     chain = parent_chain(frag)
@@ -693,15 +690,41 @@ See [Wikipedia](https://en.wikipedia.org/wiki/File:Protein_backbone_PhiPsiOmega_
 end
 
 """
+    get_torsion_psi(frag::Fragment) -> T
+Calculates the torsion angle psi (in radians) defined by the given `frag::Fragment` object.
+Returns the calculated torsion angle, if the angle can be calculated, else zero.
+The psi angle is defined by the fragment's atoms N, CA, C, and the next fragment's N atom.
+"""
+function get_torsion_psi(frag::Fragment)
+    psi = 0.0
+    if has_torsion_psi(frag)
+        next = get_next(frag)
+
+        if isnothing(next)
+            @warn "Cannot calculate psi angle, no previous or next fragment found"
+            return psi
+        end
+
+        atm_N = atom_by_name(frag, "N")
+        atm_CA = atom_by_name(frag, "CA")
+        atm_C = atom_by_name(frag, "C")
+        atm_next_N = atom_by_name(next, "N")
+
+        if isnothing(atm_N) || isnothing(atm_CA) || isnothing(atm_C) || isnothing(atm_next_N)
+            @warn "Cannot calculate psi angle, one or more atoms not found"
+            return psi
+        end
+        psi = calculate_torsion_angle(atm_N, atm_CA, atm_C, atm_next_N)
+    end
+    return psi
+end
+
+
+"""
     has_torsion_phi(frag::Fragment) -> Bool
 
-Checks if the given `Fragment` object has a torsion angle phi defined.
+Checks if the given `frag::Fragment` object has a torsion angle phi defined, `false` otherwise.
 See [Wikipedia](https://en.wikipedia.org/wiki/File:Protein_backbone_PhiPsiOmega_drawing.svg) for more information.
-# Arguments
-- `frag::Fragment`: The fragment object to be checked.
-
-# Returns
-- `Bool`: `true` if the fragment has a torsion angle phi, `false` otherwise.
 """
 @inline function has_torsion_phi(frag::Fragment)
     chain = parent_chain(frag)
@@ -712,9 +735,40 @@ See [Wikipedia](https://en.wikipedia.org/wiki/File:Protein_backbone_PhiPsiOmega_
 end
 
 """
+    get_torsion_phi(frag::Fragment) -> T
+Calculates the torsion angle phi (in radians) defined by the given `frag::Fragment` object.
+Returns the calculated torsion angle, if the angle can be calculated, else zero.
+The phi angle is defined by the fragment's atoms N, CA, C, and the previous fragment's C atom.
[…]
