### 2025-04-30 199d811
fix error in test cases due to not exporting calculate_torsion angle function and added docstrings to index
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
