+++ test/core/test_fragment.jl
@@ -1524,24 +1524,24 @@ end
         Bond(c, d, BondOrder.Single)
 
 
-        @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 180.0, atol=10e-3)
+        @test isapprox(rad2deg(BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d)), 180.0, atol=10e-3)
         @test set_torsion_angle!(a, b, c, d, T(deg2rad(90.0)))
-        @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 90.000, atol=10e-3)
+        @test isapprox(rad2deg(BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d)), 90.000, atol=10e-3)
         @test set_torsion_angle!(a, b, c, d, T(deg2rad(0.0)))
-        @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 0.0, atol=10e-3)
+        @test isapprox(rad2deg(BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d)), 0.0, atol=10e-3)
 
         @test set_torsion_angle!(a, b, c, d, T(deg2rad(149.99)))
-        @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 149.999, atol=10e-3)
+        @test isapprox(rad2deg(BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d)), 149.999, atol=10e-3)
         @test set_torsion_angle!(a, b, c, d, T(deg2rad(180.00)))
-        @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 180.0, atol=10e-3)
+        @test isapprox(rad2deg(BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d)), 180.0, atol=10e-3)
         @test set_torsion_angle!(a, b, c, d, T(deg2rad(360.00)))
-        @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 0.0, atol=10e-3)
+        @test isapprox(rad2deg(BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d)), 0.0, atol=10e-3)
         @test set_torsion_angle!(a, b, c, d, T(deg2rad(-180.00)))
-        @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 180.0, atol=10e-3)
+        @test isapprox(rad2deg(BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d)), 180.0, atol=10e-3)
         @test set_torsion_angle!(a, b, c, d, T(deg2rad(-360.00)))
-        @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), 0.0, atol=10e-3)
+        @test isapprox(rad2deg(BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d)), 0.0, atol=10e-3)
         @test set_torsion_angle!(a, b, c, d, T(deg2rad(-149.99)))
-        @test isapprox(rad2deg(calculate_torsion_angle(a, b, c, d)), -149.999, atol=10e-3)
+        @test isapprox(rad2deg(BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d)), -149.999, atol=10e-3)
 
         e = Atom(sys, 5, Elements.C, r = Vector3{T}(1.0, -1.0, 1.0))
         Bond(b, e, BondOrder.Single)
