+++ docs/src/private/fileformats.md
@@ -1,0 +1,12 @@
+# Mappings
+```@meta
+CurrentModule = BiochemicalAlgorithms
+```
+
+```@index
+Pages = ["fileformats.md"]
+```
+
+```@docs
+PDBDetails._atom_by_number
+```
