### 2025-04-30 b16db0d
add docstring
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
