### 2025-04-30 f6b45c7
Apply suggestions from code review
Co-authored-by: Thomas Kemmer <thomas@binaryden.de>
Signed-off-by: jeleclaire <leclaire@uni-mainz.de>
