+++ tutorials/getting_started.qmd
@@ -57,8 +57,8 @@ build_bonds!(s, fdb)
 
 ## How to go on?
 
-For an in depth view on the components forming the core of BiochemicalAlgorithms, we highly recommend taking a look at the next section [02_CoreComponents.md](02_CoreComponents.md) which forms a basis for the remaining tutorials dealing with different topics:
-- [03 Iterate](03_Iterate.md): How to iterate and filter over specific entities (atoms, bonds, ...)
-- [04 ForceField](04_ForceFields.md): How to use a force field and minimize a structure
-- [05 Handling Molecules](05_HandleMolecules.md): Mixed code snippets for common use cases
-- [06 Read And Write](06_ReadAndWrite.md): Read and write different file formats
+For an in depth view on the components forming the core of BiochemicalAlgorithms, we highly recommend taking a look at the next section [core_components.md](core_components.md) which forms a basis for the remaining tutorials dealing with different topics:
+- [iterate](iterate.md): How to iterate and filter over specific entities (atoms, bonds, ...)
+- [force_field_ops](force_field_ops.md): How to use a force field and minimize a structure
+- [handling_molecules](handle_molecules.md): Mixed code snippets for common use cases
+- [read_and_write](read_and_write.md): Read and write different file formats
