+++ src/core/fragment.jl
@@ -3,11 +3,19 @@ export
     FragmentTable,
     Nucleotide,
     Residue,
+    apply_torsion_angle!,
+    calculate_bond_angle,
     fragment_by_idx,
     fragments,
     get_full_name,
     get_previous,
     get_next,
+    get_torsion_omega,
+    get_torsion_phi,
+    get_torsion_psi,
+    has_torsion_omega,
+    has_torsion_phi,
+    has_torsion_psi,
     is_3_prime,
     is_5_prime,
     is_amino_acid,
@@ -666,6 +674,261 @@ end
     @assert is_amino_acid(frag)
     one_letter_code(frag.name)
 end
+
+"""
+    has_torsion_psi(frag::Fragment) -> Bool
+
+Checks if the given `frag::Fragment` object has a torsion angle psi defined, `false` otherwise.
+See [Wikipedia](https://en.wikipedia.org/wiki/File:Protein_backbone_PhiPsiOmega_drawing.svg) for more information.
+"""
+@inline function has_torsion_psi(frag::Fragment)
+    chain = parent_chain(frag)
+    if nresidues(chain) < 2
+        return false
+    end
+    return !has_flag(frag, :C_TERMINAL) && is_amino_acid(frag)
+end
+
+"""
+    get_torsion_psi(frag::Fragment{T}) -> T
+
+Calculates the torsion angle psi (in radians) defined by the given `frag::Fragment` object.
+Returns the calculated torsion angle, if the angle can be calculated, else zero.
+The psi angle is defined by the fragment's atoms N, CA, C, and the next fragment's N atom.
+"""
+function get_torsion_psi(frag::Fragment{T}) where T
+    psi = zero(T)
+    if has_torsion_psi(frag)
+        next = get_next(frag)
+
+        if isnothing(next)
+            @warn "Cannot calculate psi angle, no previous or next fragment found"
+            return psi
+        end
+
+        atm_N = atom_by_name(frag, "N")
+        atm_CA = atom_by_name(frag, "CA")
+        atm_C = atom_by_name(frag, "C")
+        atm_next_N = atom_by_name(next, "N")
+
+        if isnothing(atm_N) || isnothing(atm_CA) || isnothing(atm_C) || isnothing(atm_next_N)
+            @warn "Cannot calculate psi angle, one or more atoms not found"
+            return psi
+        end
+        psi = calculate_torsion_angle(atm_N, atm_CA, atm_C, atm_next_N)
+    end
+    return psi
+end
+
+"""
+    has_torsion_phi(frag::Fragment) -> Bool
+
+Checks if the given `frag::Fragment` object has a torsion angle phi defined, `false` otherwise.
+See [Wikipedia](https://en.wikipedia.org/wiki/File:Protein_backbone_PhiPsiOmega_drawing.svg) for more information.
+"""
+@inline function has_torsion_phi(frag::Fragment)
+    chain = parent_chain(frag)
+    if nresidues(chain) < 2
+        return false
+    end
+    return !has_flag(frag, :N_TERMINAL) && is_amino_acid(frag)
+end
+
+"""
+    get_torsion_phi(frag::Fragment{T}) -> T
+
+Calculates the torsion angle phi (in radians) defined by the given `frag::Fragment` object.
+Returns the calculated torsion angle, if the angle can be calculated, else zero.
+The phi angle is defined by the fragment's atoms N, CA, C, and the previous fragment's C atom.
+"""
+function get_torsion_phi(frag::Fragment{T}) where T
+    phi = zero(T)
+    if has_torsion_phi(frag)
+        prev = get_previous(frag)
+
+        if isnothing(prev)
+            @warn "Cannot calculate phi angle, no previous or next fragment found"
+            return phi
[…]
