### 2025-05-06 54fafaa
Functionalities for the computation and setting of torsion angles for fragments (#181)
* draft for torsion angles

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* finished set_torsion and added missing use cases

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* additional tests and missing docs added

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* Apply suggestions from code review

Co-authored-by: Thomas Kemmer <thomas@binaryden.de>
Signed-off-by: jeleclaire <leclaire@uni-mainz.de>

* added missing documentation, test cases for return types and fix of errors

Signed
[…]
