+++ test/core/test_fragment.jl
@@ -1417,3 +1417,138 @@ end
         @test nbonds(res) == nbonds(sys, fragment_idx = res.idx)
     end
 end
+
+@testitem "Fragment/*_torsion" begin
+    for T in [Float32, Float64]
+        sys = load_pdb(ball_data_path("../test/data/AlaAla.pdb"), T)
+        fdb = FragmentDB{T}()
+
+        build_bonds!(sys, fdb)
+        r1 = residues(sys)[1]
+        r2 = residues(sys)[2]
+
+        @test has_torsion_psi(r1) == true
+        @test has_torsion_psi(r2) == true
+
+        @test has_torsion_phi(r1) == true
+        @test has_torsion_phi(r2) == true
+        @test has_torsion_omega(r1) == true
+        @test has_torsion_omega(r2) == true
+
+        @test isapprox(rad2deg(get_torsion_psi(r1)), -57.97745f0, atol=10e-3)
+        @test isapprox(rad2deg(get_torsion_psi(r2)), -0.0, atol=10e-3)
+
+        @test isapprox(rad2deg(get_torsion_phi(r1)), 0.0, atol=10e-3)
+        @test isapprox(rad2deg(get_torsion_phi(r2)), -47.03374, atol=10e-3)
+
+        @test isapprox(rad2deg(get_torsion_omega(r1)), 179.9903002, atol=10e-3)
+        @test isapprox(rad2deg(get_torsion_omega(r2)), 0.0, atol=10e-3)
+
+        sys2 = deepcopy(sys)
+        @test nfragments(sys2) == 2
+        @test nresidues(sys2) == 2
+        delete!(residues(sys2)[2])
+        @test nresidues(sys2) == 1
+
+        r1 = residues(sys2)[1]
+        @test has_torsion_psi(r1) == false
+        @test has_torsion_phi(r1) == false
+        @test has_torsion_omega(r1) == false
+
+        r1 = residues(sys)[1]
+        r2 = residues(sys)[2]
+
+        @test !has_flag(r1, :N_TERMINAL)
+        @test !has_flag(r1, :C_TERMINAL)
+
+        @test !has_flag(r2, :N_TERMINAL)
+        @test !has_flag(r2, :C_TERMINAL)
+
+        set_flag!(r1, :N_TERMINAL)
+        @test has_flag(r1, :N_TERMINAL)
+
+        set_flag!(r2, :C_TERMINAL)
+        @test has_flag(r2, :C_TERMINAL)
+
+        @test has_torsion_psi(r1) == true
+        @test has_torsion_psi(r2) == false
+
+        @test has_torsion_phi(r1) == false
+        @test has_torsion_phi(r2) == true
+        @test has_torsion_omega(r1) == false
+        @test has_torsion_omega(r2) == true
+    end
+end
+
+
+@testitem "Fragment/calculate_torsion_angle" begin
+    for T in [Float32, Float64]
+        sys = System{T}()
+        a = Atom(sys, 1, Elements.C, r = Vector3{T}(0.0f0, 1.0f0, 0.0f0))
+        b = Atom(sys, 2, Elements.C, r = Vector3{T}(0.0f0, 1.0f0, -1.0f0))
+        c = Atom(sys, 3, Elements.N, r = Vector3{T}(1.0f0, 0.0f0, 0.0f0))
+        d = Atom(sys, 4, Elements.C, r = Vector3{T}(1.0f0, -1.0f0, -1.0f0))
+        @test BiochemicalAlgorithms.calculate_torsion_angle(a, b, c, d) isa T
[…]
