### 2025-05-07 ed2663f
Apply suggestions from code review
Co-authored-by: Thomas Kemmer <thomas@binaryden.de>
Signed-off-by: jeleclaire <leclaire@uni-mainz.de>
