### 2025-06-02 5815a6d
CompatHelper: bump compat for MolecularGraph to 0.19, (keep existing compat)
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
