### 2025-06-13 719b6ca
FF: fix parameter assignment on Julia nightly
Fixes: #192
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
