### 2025-06-13 0f363ab
FF: fix parameter assignment on Julia nightly
Fixes: #192
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
