### 2025-06-17 #174 @fcdimitr
Hi @okohlbacher 👋 ,

would you like to review this JuliaCon proceedings submission?

you can read more about reviewing for JuliaCon proceedings in the [reviewer guide](https://juliacon.github.io/proceedings-guide/reviewer/).
