### 2025-06-22 bc86ded
CompatHelper: bump compat for MolecularGraph to 0.20
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
