### 2025-06-23 94caee3
FragmentDB: create bonds in fragment insertion order
This helps keeping the atom ordering in bonds consistent between Julia
versions (esp. before and after Julia 1.13 due to hashing changes)

Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
