### 2025-06-23 8b63ca1
FragmentDB: create bonds in fragment insertion order
This helps keeping the atom ordering in bonds consistent between Julia
versions (esp. before and after Julia 1.13 due to hashing changes)

Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
