### 2025-06-24 f240d51
CompatHelper: bump compat for MolecularGraph to 0.20, (keep existing compat)
