+++ Project.toml
@@ -53,7 +53,7 @@ JSON3 = "1.9.1"
 LinearAlgebra = "1.10"
 Mendeleev = "1"
 MetaGraphs = "0.8"
-MolecularGraph = "0.16, 0.17, 0.18, 0.19.1"
+MolecularGraph = "0.16, 0.17, 0.18, 0.19.1, 0.20"
 Observables = "0.5.5"
 Optimization = "4"
 PhysicalConstants = "0.2.3"
