### 2025-06-24 717592e
Migrate to MolecularGraph.jl v0.20
Closes: #197
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
