### 2025-06-24 6aabe44
v0.5.4
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
