+++ v0.5.4/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"public/forcefields/#Force-fields","page":"Force fields","title":"Force fields","text":"","category":"section"},{"location":"public/forcefields/","page":"Force fields","title":"Force fields","text":"Pages = [\"forcefields.md\"]","category":"page"},{"location":"public/forcefields/#BiochemicalAlgorithms.AmberFF","page":"Force fields","title":"BiochemicalAlgorithms.AmberFF","text":"AmberFF(\n    ::AbstractAtomContainer{T},\n    param_file::AbstractString = ball_data_path(\"forcefields/AMBER/amber96.ini\")\n)\n\nInitializes an AMBER force field for the given atom container and the given parameter file (default: AMBER96).\n\nSupported keyword arguments\n\nnonbonded_cutoff::T = 20\nvdw_cutoff::T = 15\nvdw_cuton::T = 13\nelectrostatic_cutoff::T = 15\nelectrostatic_cuton::T = 13\nscaling_vdw_1_4 = 2\nscaling_electrostatic_1_4::T = 1.2\ndistance_dependent_dielectric::Bool = false\nassign_charges::Bool = true\nassign_typenames::Bool = true\nassign_types::Bool = true\noverwrite_nonzero_charges::Bool = true\noverwrite_typenames::Bool = false\nperiodic_boundary_conditions::Bool = false\nperiodic_box_width::T = 100\nperiodic_box_height::T = 100\nperiodic_box_depth::T = 100\nmax_number_of_unassigned_atoms::Int = typemax(Int32)\n\n\n\n\n\n","category":"function"},{"location":"public/forcefields/#BiochemicalAlgorithms.optimize_hydrogen_positions!","page":"Force fields","title":"BiochemicalAlgorithms.optimize_hydrogen_positions!","text":"optimize_hydrogen_positions!(ff::ForceField)\n\nVariant of optimize_structure! that only optimizes hydrogen atom positions.\n\nSupported keyword arguments\n\nSame as optimize_structure!\n\n\n\n\n\n","category":"function"},{"location":"public/forcefields/#BiochemicalAlgorithms.optimize_structure!","page":"Force fields","title":"BiochemicalAlgorithms.optimize_structure!","text":"optimize_structure!(ff::ForceField)\n\nAttempts to solve the energy optimization problem represented by the given force field object.\n\nSupported keyword arguments\n\nThis function passes all keyword arguments to Optimization.solve, with the following default values:\n\nalg = Optimization.LBFGS()\n\n\n\n\n\n","category":"function"},{"location":"public/forcefields/#BiochemicalAlgorithms.update!","page":"Force fields","title":"BiochemicalAlgorithms.update!","text":"update!(ff::ForceField)\n\n\nUpdate the internal data structures of the force field when the system changes (e.g., through coordinate updates).\n\nnote: Note\nChanges to the options or the topology require a call to setup! prior to the call to update!.\n\n\n\n\n\n","category":"function"},{"location":"public/mappings/#Mappings","page":"Mappings","title":"Mappings","text":"","category":"section"},{"location":"public/mappings/","page":"Mappings","title":"Mappings","text":"Pages = [\"mappings.md\"]","category":"page"},{"location":"public/mappings/#Atom-bijections","page":"Mappings","title":"Atom bijections","text":"","category":"section"},{"location":"public/mappings/#BiochemicalAlgorithms.AbstractAtomBijection","page":"Mappings","title":"BiochemicalAlgorithms.AbstractAtomBijection","text":"abstract type AbstractAtomBijection{T}\n\nAbstract base 
[…]
