### 2025-07-07 07e700f
Update compat lower bounds for MolecularGraph v0.20
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
