+++ Project.toml
@@ -48,8 +48,8 @@ DataStructures = "0.18.13"
 Dates = "1.10"
 DocStringExtensions = "0.9"
 EnumX = "1"
-Graphs = "1.8"
-JSON3 = "1.9.1"
+Graphs = "1.12"
+JSON3 = "1.10"
 LinearAlgebra = "1.10"
 Mendeleev = "1"
 MetaGraphs = "0.8"
@@ -64,7 +64,7 @@ Rotations = "1.3.4"
 Scanf = "0.5.4"
 StaticArrays = "1.5"
 Statistics = "1.10"
-StructTypes = "1.5"
+StructTypes = "1.10"
 TableOperations = "1.2"
 Tables = "1.11"
 Test = "1.10"
