### 2025-07-07 164592b
Update Aqua compat lower bound for Julia nightly
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
