### 2025-07-07 f5efa5a
CI: also trigger Downgrade worfklow on changes to Project.toml
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
