+++ .github/workflows/Downgrade.yml
@@ -1,5 +1,8 @@
 name: CompatDowngrade
 on:
+  push:
+    paths:
+      - 'Project.toml'
   schedule:
     - cron: 0 0 * * *
   workflow_dispatch:
