+++ Project.toml
@@ -44,7 +44,7 @@ BioSymbols = "5"
 CSV = "0.10.13"
 CellListMap = "0.8, 0.9"
 DataFrames = "1.6"
-DataStructures = "0.18.13"
+DataStructures = "0.18.13, 0.19"
 Dates = "1.10"
 DocStringExtensions = "0.9"
 EnumX = "1"
