### 2025-08-02 ddf2b6e
CompatHelper: bump compat for DataStructures to 0.19, (keep existing compat)
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
