+++ Project.toml
@@ -42,7 +42,7 @@ BioSymbols = "5"
 CSV = "0.10.13"
 CellListMap = "0.8, 0.9"
 DataFrames = "1.6"
-DataStructures = "0.18.13"
+DataStructures = "0.18.13, 0.19"
 DocStringExtensions = "0.9"
 EnumX = "1"
 Graphs = "1.8"
