+++ .github/workflows/DCO.yml
@@ -5,7 +5,7 @@ jobs:
   dco-check:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
+    - uses: actions/checkout@v5
     - name: Set up Python 3.x
       uses: actions/setup-python@v5
       with:
