### 2025-08-18 61d2f0a
CI: exclude dependabot from DCO checks
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
