### 2025-08-18 4fd83df
Compat: bump Aqua compat lower bound for Julia nightly
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
