+++ Project.toml
@@ -37,7 +37,7 @@ Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 UnitfulAtomic = "a7773ee8-282e-5fa2-be4e-bd808c38a91a"
 
 [compat]
-Aqua = "0.8.13"
+Aqua = "0.8.14"
 AutoHashEquals = "2"
 BioStructures = "4.2"
 BioSymbols = "5"
