### 2025-08-25 4ce23f8
v0.5.5
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
