### 2025-08-25 132e8a6
Compat: bump DataFrames compat lower bound to skip yanked version
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
