### 2025-08-25 559d7ae
Compat: bump CellListMap compat lower bound to 0.9.5
Earlier versions don't support TestItems v1, which causes
precompilation errors with TestItemRunner.

Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
