+++ Project.toml
@@ -41,7 +41,7 @@ AutoHashEquals = "2"
 BioStructures = "4.2"
 BioSymbols = "5"
 CSV = "0.10.13"
-CellListMap = "0.8, 0.9"
+CellListMap = "0.9.5"
 DataFrames = "1.6"
 DataStructures = "0.18.13, 0.19"
 DocStringExtensions = "0.9"
