+++ Project.toml
@@ -42,7 +42,7 @@ AutoHashEquals = "2"
 BioStructures = "4.2"
 BioSymbols = "5"
 CSV = "0.10.13"
-CellListMap = "0.8, 0.9"
+CellListMap = "0.9.5"
 DataFrames = "1.6.1"
 DataStructures = "0.18.13, 0.19"
 Dates = "1.10"
