### 2025-08-26 63d00ef
CompatHelper: bump compat for PrettyTables to 3
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
