+++ Project.toml
@@ -43,7 +43,7 @@ BioStructures = "4.2"
 BioSymbols = "5"
 CSV = "0.10.13"
 CellListMap = "0.9.5"
-DataFrames = "1.6.1"
+DataFrames = "1.8"
 DataStructures = "0.18.13, 0.19"
 Dates = "1.10"
 DocStringExtensions = "0.9"
@@ -57,7 +57,7 @@ MolecularGraph = "0.20.2, 0.21"
 Observables = "0.5.5"
 Optimization = "4"
 PhysicalConstants = "0.2.3"
-PrettyTables = "2.2.7"
+PrettyTables = "3"
 Printf = "1.10"
 Quaternions = "0.7.6"
 Rotations = "1.3.4"
