### 2025-08-26 84f3b7e
CompatHelper: bump compat for PrettyTables to 3, (keep existing compat)
