+++ Project.toml
@@ -57,7 +57,7 @@ MolecularGraph = "0.20.2"
 Observables = "0.5.5"
 Optimization = "4"
 PhysicalConstants = "0.2.3"
-PrettyTables = "2.2.7"
+PrettyTables = "2.2.7, 3"
 Printf = "1.10"
 Quaternions = "0.7.6"
 Rotations = "1.3.4"
