### 2025-08-28 3ed86dd
CompatHelper: bump compat for MolecularGraph to 0.21, (keep existing compat)
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
