+++ Project.toml
@@ -53,7 +53,7 @@ JSON3 = "1.10"
 LinearAlgebra = "1.10"
 Mendeleev = "1"
 MetaGraphs = "0.8"
-MolecularGraph = "0.20.2"
+MolecularGraph = "0.20.2, 0.21"
 Observables = "0.5.5"
 Optimization = "4"
 PhysicalConstants = "0.2.3"
