### 2025-09-01 6260971
added missing section about secondary structure in core tutorial
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
