### 2025-09-04 cec87f7
MolecularGraphWrapper: ignore atom mass
Handling of atoms masses has changed in recent versions of
MolecularGraphs and this information is mostly unused and
completely untested in our code.

Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
