### 2025-09-04 e303814
CI: don't force latest compatible version for downgrade checks
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
