+++ .github/workflows/DCO.yml
@@ -7,7 +7,7 @@ jobs:
     steps:
     - uses: actions/checkout@v5
     - name: Set up Python 3.x
-      uses: actions/setup-python@v5
+      uses: actions/setup-python@v6
       with:
         python-version: '3.x'
     - name: Check DCO
