### 2025-09-11 b623119
System: migrate to PrettyTables v3
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
