### 2025-09-12 e93e334
include section for secondary structure from pdb
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
