+++ docs/src/tutorials/core_components.md
@@ -16,12 +16,12 @@ s = load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
 s
 ```
 
-    System{Float32}: AlaAla.pdb
+    System{Float32}: AA
       23 atoms
        0 bonds
        1 molecules
        1 chains
-       0 secondary structures
+       1 secondary structures
        2 fragments
 
 This will give us an overview of the objects associated with our system. It contains 23 atoms, 0 bonds, 1 molecule, 1 chain, 0 secondary structures and 2 fragments. We look at each group separately in the following sections.
@@ -32,12 +32,12 @@ If we do not explicitely create a system, the data will be stored in the default
 load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
 ```
 
-    System{Float32}: AlaAla.pdb
+    System{Float32}: AA
       23 atoms
        0 bonds
        1 molecules
        1 chains
-       0 secondary structures
+       1 secondary structures
        2 fragments
 
 ## Atom
@@ -61,14 +61,14 @@ for a in sys_atms
 end
 println()
 
-n1= atom_by_name(s, "N")
+n1 = atom_by_name(s, "N")
 
 println(n1.idx)
 ```
 
     Atom elements:
     NCCOHHHHHHCHNCCHHHHCHOO
-    5
+    4
 
 We can access an atom by its name, which will return the first atom matching the given name. If there are more atoms with the same name and we are not interested in the first one, we can access the atom by the atom index.
 The atom index is unique inside a system but not necessarily starting by one.
@@ -83,18 +83,18 @@ n2 = atom_by_idx(s, 17)
 ```
 
     Atom ids:
-    5 N
+    4 N
+    5 C
     6 C
-    7 C
-    8 O
+    7 O
+    8 H
     9 H
     10 H
     11 H
     12 H
     13 H
-    14 H
-    15 C
-    16 H
+    14 C
+    15 H
     17 N
     18 C
     19 C
@@ -117,18 +117,18 @@ for a in sys_atms
 end
 ```
 
-    5 1
-    6 2
-    7 3
-    8 4
-    9 5
-    10 6
-    11 7
-    12 8
-    13 9
-    14 10
-    15 11
-    16 12
+    4 1
+    5 2
+    6 3
+    7 4
+    8 5
+    9 6
+    10 7
+    11 8
+    12 9
+    13 10
+    14 11
+    15 12
     17 13
     18 14
     19 15
@@ -178,21 +178,10 @@ using Pkg
 Pkg.add("BiochemicalVisualization")
 ```
 
-       Resolving package versions...
-          Compat entries added for BiochemicalVisualization
-        Updating `~/local/BiochemicalAlgorithms.jl/Project.toml`
-      [00dfb448] + BiochemicalVisualization v0.3.2
-      No Changes to `~/local/BiochemicalAlgorithms.jl/Manifest.toml`
-    ┌ Warning: Circular dependency detected.
-    │ Precompilation will be skipped for dependencies in this cycle:
-    │  ┌ BiochemicalAlgorithms
-    │  └─ BiochemicalVisualization
-    └ @ Base.Precompilation precompilation.jl:650
-
 ``` julia
 using BiochemicalVisualization
 
-# ball_and_stick(s)  uncomment for visualization
+ball_and_stick(s)
 ```
 
 Now, let’s create some bonds:
@@ -218,7 +207,7 @@ for i in residues(s)
 end
 ```
 
[…]
