+++ docs/src/tutorials/handle_molecules.md
@@ -134,14 +134,23 @@ end
 
     > Chain E
     One-letter code:
-    IVGGYTCGANTVPYQVSLNSGYHFCGGSLINSQWVVSAAHCYKSGIQVRLGEDNINVVEGNEQFISASKSIVHPSYNSNTLNNDIMLIKLKSAASLNSRVASISLPTSCASAGTQCLISGWGNTKSSGTSYPDVLKCLKAPILSDSSCKSAYPGQITSNMFCAYGLEGKGDSCQGDSGGPVVCSGKLQGIVSWGSGCQAKNKPGVYTKVCNYVSWIKQTIASN
+    IVGGYTCGANTVPYQVSLNSGYHFCGGSLINSQWVVSAAHCYKSGIQVRLGEDNINVVEGNEQFISASKSIVHPSYNSNTLNNDIMLIKLKSAASLNSRVASISLPTSCASAGTQCLISGWGNTKSSGTSYPDVLKCLKAPILSDSSCKSAYPGQITSNMFCAGYLEGGKDSCQGDSGGPVVCSGKLQGIVSWGSGCAQKNKPGVYTKVCNYVSWIKQTIASN
     Three-letter code:
-    ILE VAL GLY GLY TYR THR CYS GLY ALA ASN THR VAL PRO TYR GLN VAL SER LEU ASN SER GLY TYR HIS PHE CYS GLY GLY SER LEU ILE ASN SER GLN TRP VAL VAL SER ALA ALA HIS CYS TYR LYS SER GLY ILE GLN VAL ARG LEU GLY GLU ASP ASN ILE ASN VAL VAL GLU GLY ASN GLU GLN PHE ILE SER ALA SER LYS SER ILE VAL HIS PRO SER TYR ASN SER ASN THR LEU ASN ASN ASP ILE MET LEU ILE LYS LEU LYS SER ALA ALA SER LEU ASN SER ARG VAL ALA SER ILE SER LEU PRO THR SER CYS ALA SER ALA GLY THR GLN CYS LEU ILE SER GLY TRP GLY ASN THR LYS SER SER GLY THR SER TYR PRO ASP VAL LEU LYS CYS LEU LYS ALA PRO ILE LEU SER ASP SER SER CYS LYS SER ALA TYR PRO GLY GLN ILE THR SER ASN MET PHE CYS ALA TYR GLY LEU GLU GLY LYS GLY ASP SER CYS GLN GLY ASP SER GLY GLY PRO VAL VAL CYS SER GLY LYS LEU GLN GLY ILE VAL SER TRP GLY SER GLY CYS GLN ALA LYS ASN LYS PRO GLY VAL TYR THR LYS VAL CYS ASN TYR VAL SER TRP ILE LYS GLN THR ILE ALA SER ASN
[…]
