+++ docs/src/tutorials/core_components.md
@@ -211,67 +211,36 @@ end
     Set([:N_TERMINAL])
     Set([:C_TERMINAL])
 
-Let’s take a deeper look at the bonds:
+Let’s take a deeper look at the bonds. Bond are only formed inside the same system and not across systems.
 
 ``` julia
-for bond in bonds(s)
-    println(bond)
-end
+bonds(s)
 ```
 
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 29, a1 = 4, a2 = 12, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 30, a1 = 4, a2 = 5, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 31, a1 = 4, a2 = 8, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 32, a1 = 4, a2 = 10, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 33, a1 = 5, a2 = 15, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 34, a1 = 5, a2 = 6, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 35, a1 = 5, a2 = 14, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 36, a1 = 6, a2 = 7, order = BiochemicalAlgorithms.BondOrder.Double)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 37, a1 = 9, a2 = 14, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 38, a1 = 11, a2 = 14, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 39, a1 = 13, a2 = 14, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 40, a1 = 17, a2 = 20, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 41, a1 = 17, a2 = 18, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 42, a1 = 18, a2 = 25, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 43, a1 = 18, a2 = 19, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 44, a1 = 18, a2 = 24, order = BiochemicalAlgorithms.BondOrder.Single)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 45, a1 = 19, a2 = 26, order = BiochemicalAlgorithms.BondOrder.Double)
-    BiochemicalAlgorithms.Bond{Float32}: (idx = 46, a1 = 19, a2 = 27, order = BiochemicalAlgorithms.BondOrder.Single)
[…]
