### 2025-09-19 514bffe
A new tutorial for the core components including secondary structure (#187)
* tutorials: added small use cases

Signed-off-by: Jenny <leclaire@uni-mainz.de>
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* working on new tutorial

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* restructuring of tutorials

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* restructuring and additional tutorial

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* removed outdated md files

Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>

* update links

Signed-off-by: Jenny Lecl
[…]
