+++ docs/src/tutorials/iterate.md
@@ -5,18 +5,18 @@ When working with molecular entities, we want to run over all atoms of a system,
 
 ## Molecular systems
 
-In `BiochemicalAlgorithms.jl` atoms and bonds are existing inside a `System`. Typically, molecular data is stored in molecular data formats such as PDB. The latter can be directly read into a system.
+In `BiochemicalAlgorithms.jl` atoms and bonds are existing inside a [`System`](@ref). Typically, molecular data is stored in molecular data formats such as PDB. The latter can be directly read into a system.
 
 ``` julia
 sys = load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
 ```
 
-    System{Float32}: AlaAla.pdb
+    System{Float32}: AA
       23 atoms
        0 bonds
        1 molecules
        1 chains
-       0 secondary structures
+       1 secondary structures
        2 fragments
 
 You can, e.g., print all atoms of the given system as a table:
@@ -27,29 +27,29 @@ atoms(sys)
 
 | **\#** | **idx** | **number** | **element** | **name** | **atom_type** | **r** | **v** | **F** | **formal_charge** | **charge** | **radius** |
 |---:|:---|:---|:---|:---|:---|:---|:---|:---|:---|:---|:---|
-| **1** | 5 | 1 | N | N |  | Float32\[-1.45, 0.0, 0.0\] | Float32\[0.0, 0.0, 0.0\] | Float32\[0.0, 0.0, 0.0\] | 0 | 0.0 | 0.0 |
-| **2** | 6 | 2 | C | CA |  | Float32\[0.0, 0.0, 0.0\] | Float32\[0.0, 0.0, 0.0\] | Float32\[0.0, 0.0, 0.0\] | 0 | 0.0 | 0.0 |
-| **3** | 7 | 3 | C | C |  | Float32\[0.495, 0.0, 1.437\] | Float32\[0.0, 0.0, 0.0\] | Float32\[0.0, 0.0, 0.0\] | 0 | 0.0 | 0.0 |
-| **4** | 8 | 4 | O | O |  | Float32\[1.235, -0.911, 1.838\] | Float32\[0.0, 0.0, 0.0\] | Float32\[0.0, 0.0, 0.0\] | 0 | 0.0 | 0.0 |
-| **5** | 9 | 5 | H | 1H |  | Float32\[-1.788, 0.918, 0.25\] | Float32\[0.0, 0.0, 0.0\] | Float32\[0.0, 0.0, 0.0\] | 0 | 0.0 | 0.0 |
-| **6** | 10 | 6 | H | 1HB |  | Float32\[1.642, 1.124, -0.864\] | Float32\[0.0, 0.0, 0.0\] | Float32\[0.0, 0.0, 0.0\] | 0 | 0.0 | 0.0 |
-| **7** | 11 | 7 | H | 2H |  | Float32\[-1.801, -0.26, -0.911\] | Float32\[0.0, 0.0, 0.0\] | Float32\[0.0, 0.0, 0.0\] | 0 | 0.0 | 0.0 |
[…]
