+++ docs/src/tutorials/core_components.md
@@ -1,0 +1,610 @@
+# Getting started
+
+
+BiochemicalAlgorithms.jl\`s design is inspired by its predecessor [`BALL`](https://github.com/ball-project/ball). BALL and BiochemicalAlgorithms.jl both have a set of KERNEL classes or core classes to represent molecular entities such as Molecules, Proteins, Atoms, Bonds etc. The following class diagramm shows the different classes and their relationship to each others.
+
+![UML class diagram *Note*: Only the most important functionalities are shown here.](uml.png)
+
+UML class diagram *Note*: Only the most important functionalities are shown here.
+
+In the center of all classes resides the [`System`](@ref). So let’s see it in action with a simple peptide:
+
+``` julia
+using BiochemicalAlgorithms
+
+s = load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
+s
+```
+
+    System{Float32}: AA
+      23 atoms
+       0 bonds
+       1 molecules
+       1 chains
+       1 secondary structures
+       2 fragments
+
+This will give us an overview of the objects associated with our system. It contains 23 atoms, 0 bonds, 1 molecule, 1 chain, 0 secondary structures and 2 fragments. We look at each group separately in the following sections.
+
+If we do not explicitely create a system, the data will be stored in the default system:
+
+``` julia
+load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
+```
+
+    System{Float32}: AA
+      23 atoms
+       0 bonds
+       1 molecules
+       1 chains
+       1 secondary structures
+       2 fragments
+
+## Atom
+
+We can access the atoms like this:
+
+``` julia
+sys_atms = atoms(s)
+
+natoms(sys_atms)
+```
+
+    23
+
+Let’s play around with it:
+
+``` julia
+println("Atom elements:")
+for a in sys_atms
+    print(a.element)
+end
+println()
+
+n1 = atom_by_name(s, "N")
+
+println(n1.idx)
+```
+
+    Atom elements:
+    NCCOHHHHHHCHNCCHHHHCHOO
+    4
+
+We can access an atom by its name, which will return the first atom matching the given name. If there are more atoms with the same name and we are not interested in the first one, we can access the atom by the atom index.
+The atom index is unique inside a system but not necessarily starting by one.
+
+``` julia
+println("Atom ids:")
+for a in sys_atms
+    println(a.idx, " ", a.element)
+end
+
+n2 = atom_by_idx(s, 17)
+```
+
+    Atom ids:
+    4 N
+    5 C
+    6 C
+    7 O
+    8 H
+    9 H
+    10 H
+    11 H
+    12 H
+    13 H
+    14 C
+    15 H
+    17 N
+    18 C
+    19 C
+    20 H
+    21 H
+    22 H
+    23 H
+    24 C
+    25 H
+    26 O
+    27 O
+
+    BiochemicalAlgorithms.Atom{Float32}: (idx = 17, number = 13, element = BiochemicalAlgorithms.Elements.N, name = "N", atom_type = "", r = Float32[0.133, 0.98, 2.268], v = Float32[0.0, 0.0, 0.0], F = Float32[0.0, 0.0, 0.0], formal_charge = 0, charge = 0.0f0, radius = 0.0f0)
+
+[`Atom`](@ref) features the atom number as well. The atom number does NOT have to be unique inside the system!
+
+``` julia
+for a in sys_atms
+    println(a.idx, " ", a.number)
+end
+```
+
+    4 1
+    5 2
+    6 3
+    7 4
+    8 5
+    9 6
+    10 7
+    11 8
+    12 9
+    13 10
+    14 11
+    15 12
+    17 13
[…]
