### 2025-09-30 #174 @fcdimitr
Hi @michielstock 👋 ,

I'm sorry for pinging again. Would you like to review this JuliaCon proceedings submission?

you can read more about reviewing for JuliaCon proceedings in the [reviewer guide](https://juliacon.github.io/proceedings-guide/reviewer/).
