### 2025-09-30 #188 @kescobo
- [x] I confirm that I have read the [JuliaCon conflict of interest policy](https://juliacon.github.io/proceedings-guide/reviewer/#conflict_of_interest) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JCon for the purpose of this review.

- [x] **Authors:** Does the `paper.pdf` file include a list of authors with their affiliations?
- [x] **A statement of need:** Does the paper clearly explain the problem it addresses, its importance to the broader community, the intended audience, and how it connects to existing work in the field?
