### 2025-10-15 301c408
CompatHelper: bump compat for Optimization to 5.1
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
