+++ Project.toml
@@ -55,7 +55,7 @@ Mendeleev = "1"
 MetaGraphs = "0.8"
 MolecularGraph = "0.20.2, 0.21"
 Observables = "0.5.5"
-Optimization = "4"
+Optimization = "5.1"
 PhysicalConstants = "0.2.3"
 PrettyTables = "3"
 Printf = "1.10"
