### 2025-10-15 7a9422f
Optim: adapt to Optimization.jl v5
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
