### 2025-10-20 6af8a5e
CompatHelper: bump compat for MolecularGraph to 0.22
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
