+++ Project.toml
@@ -54,7 +54,7 @@ JSON3 = "1.10"
 LinearAlgebra = "1.10"
 Mendeleev = "1"
 MetaGraphs = "0.8"
-MolecularGraph = "0.20.2, 0.21"
+MolecularGraph = "0.22"
 Observables = "0.5.5"
 Optimization = "5.1"
 OptimizationLBFGSB = "1.1.0"
