### 2025-10-28 hildebrandtlab/BiochemicalAlgorithms.jl#212 @jeleclaire
Example code in tutorial needs update for setting atom positon upon construction of atoms
As described by [vboussange in his review](https://github.com/JuliaCon/proceedings-review/issues/188#issuecomment-3410648720) assigning the positions of the atoms after construction is not straightforward.  We can already assign the positions in the constructor. The referenced example in the tutorials section just needs to be updated accordingly.
