### 2025-10-28 c7c38b8
Update constructor calls of atoms to set atom position (closes  #212)
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
