### 2025-10-28 #188 @kescobo
Just went through, and it looks good! Agree with @vboussange that this is an excellent example of julia's strengths, andi like the direct comparisons with the BALL C++ code. Also agree re: test, mutable structs, and DOIs. In addition:f

1. The `ball_data_path()` function is used all over the place in the docs and tutorials but doesn't have a docstring or any description when it's introduced. I recognize that it's just a helper function to make your lives easier for these things, but I think some explanation is warranted
2. While I recognize that this is not directly about your visualization package, since you talk about it in the paper, it might be nice to have a few sentences on pros/cons relative to BioMakie.jl
