+++ paper/ref.bib
@@ -113,16 +113,17 @@
 }
 
 @article{Greener2024,
-	author = {Greener, Joe G},
-	title = {Differentiable simulation to develop molecular dynamics force fields for disordered proteins},
+	author = {Joe G. Greener},
+	title = {Differentiable simulation to develop molecular dynamics force fields for disordered proteins††Electronic supplementary information (ESI) available. See DOI: https://doi.org/10.1039/d3sc05230c},
 	journal = {Chemical Science},
 	volume = {15},
+	number = {13},
 	pages = {4897-4909},
-	year = {2024}
+	year = {2024},
+	issn = {2041-6520},
+	doi = {https://doi.org/10.1039/d3sc05230c},
+	url = {https://www.sciencedirect.com/science/article/pii/S2041652024003845}
 }
-
-
-
 
 @misc{Julia_accomplish,
 	author = {Julia Data Science},
