### 2025-10-31 hildebrandtlab/BiochemicalAlgorithms.jl#214 @jeleclaire
Integration of JET.jl in CI Pipeline
We regularly run `report_package` from JET.jl locally to detect type instabilities. As suggested by [vboussange](https://github.com/JuliaCon/proceedings-review/issues/188#issuecomment-3410648720) we should strongly consider to integrate it into our CI Pipeline.
