### 2025-10-31 0c7ed2d
added missing DocString for ball_data_path
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
