### 2025-10-31 5c028d4
added missing DocString for ball_data_path
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
