+++ docs/src/public/utils.md
@@ -1,0 +1,4 @@
+# Utility functions
+```@docs
+ball_data_path
+```
