### 2025-10-31 hildebrandtlab/BiochemicalAlgorithms.jl#215 @jeleclaire
Benchmark `System` as immutable struct
https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl/blob/514bffe1cc5c371d74b2c849839407690aa2701a/src/core/system.jl#L109

- [ ] Implement different indexing scheme
- [ ] Run benchmarks in comparison to mutable version
