### 2025-11-04 1a3a15f
CI: replace julia-actions/julia-downgrade-compat
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
