### 2025-11-04 c1ab018
Compat: bump StaticArrays compat lower bound to 1.9.8
A breaking patch release from StaticArraysCore broke precompilation of
all previous versions of StaticArrays w/ Julia LTS:
https://github.com/JuliaArrays/StaticArrays.jl/issues/1323

Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
