+++ Project.toml
@@ -62,7 +62,7 @@ Printf = "1.10"
 Quaternions = "0.7.6"
 Rotations = "1.3.4"
 Scanf = "0.5.4"
-StaticArrays = "1.5"
+StaticArrays = "1.9.8"
 Statistics = "1.10"
 StructTypes = "1.10"
 TableOperations = "1.2"
