### 2025-11-04 bca0356
Compat: bump StaticArrays compat lower bound to 1.9.8
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
