+++ Project.toml
@@ -60,7 +60,7 @@ PrettyTables = "2.2.7"
 Printf = "1.10"
 Quaternions = "0.7.6"
 Rotations = "1.3.4"
-StaticArrays = "1.5"
+StaticArrays = "1.9.8"
 Statistics = "1.10"
 StructTypes = "1.5"
 TableOperations = "1.2"
