### 2025-11-04 hildebrandtlab/BiochemicalAlgorithms.jl#216 @tkemmer
`optimize_structure` tests fail intermittently on Github runners
```
Optimize structure: Test Failed at /home/runner/work/BiochemicalAlgorithms.jl/BiochemicalAlgorithms.jl/test/optimization/test_optimize_structure.jl:11
  Expression: Int((optimize_structure!(ff)).retcode) == 1
   Evaluated: 7 == 1

Stacktrace:
 [1] macro expansion
   @ /opt/hostedtoolcache/julia/1.10.10/x64/share/julia/stdlib/v1.10/Test/src/Test.jl:672 [inlined]
 [2] top-level scope
   @ ~/work/BiochemicalAlgorithms.jl/BiochemicalAlgorithms.jl/test/optimization/test_optimize_structure.jl:508
```
This happens on different Julia versions and with different versions of Optimization.jl. So far, this error occurred only on Github runners and we were not able to reproduce this locally.

The observed return code corresponds to `ReturnCode.Unstable`:
https://github.com/SciML/SciMLBase.jl/blob/92d846072d883c18f98cdc62c49189b833649159/src/retcodes.jl#L219
