### 2025-11-04 4d7a38e
CI: skip optimization tests for now (#216)
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
