+++ test/runtests.jl
@@ -1,3 +1,7 @@
 using TestItemRunner
 
-@run_package_tests verbose=true
+if get(ENV, "GITHUB_ACTIONS", "false") == "true"
+    @run_package_tests verbose=true filter=ti->!(:skip_ci in ti.tags)
+else
+    @run_package_tests verbose=true
+end
