### 2025-11-04 2e54cc1
CI: skip optimization tests for now (#216)
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
