### 2025-11-04 4806e58
CI: bump Julia version to 1.12, drop 1.11
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
