### 2025-11-04 hildebrandtlab/BiochemicalAlgorithms.jl#218 @tkemmer
Testing broken w/ Julia 1.13+
Upstream bug https://github.com/julia-vscode/TestItemRunner.jl/issues/113
