### 2025-11-04 df35857
CI: disable Julia nightly testing for now (#218)
Let's just not waste resources unnecessarily until this is resolved
upstream or we've decided to switch test runners.

Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
