### 2025-11-04 e8930e1
MolecularGraphWrapper: adapt to MolecularGraph 0.22
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
