### 2025-11-06 #188 @jeleclaire
Hi @vboussange and @kescobo!
Thank you for your constructive feedback and valuable suggestions.

I will summarize the addressed points:

- [x] Constructor call of `Atom` (see [PR](https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl/pull/219)):
     Both constructor calls are possible but your suggestions is more convenient. I updated the documentation accordingly.

- [x] Missing DocString for `ball_data_path` (see [PR](https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl/pull/219)):
     Thanks for pointing that out! It was indeed missing. It is now included in the documentation.

- [x] Missing DOIs see update in paper branch:
    - [x] 4 [Benchmarktools.jl](https://github.com/JuliaCI/BenchmarkTools.jl/blob/main/CITATION.bib) has no DOI
    - [x] 7 Design Patterns has no DOI
    - [x] 8 DOI added

- [x] Jet.jl:
is planned for the future. We run `report_package` from JET.jl  from time to time locally to detect type instabilities. In the past, it has tended to generate verbose outputs for our dependencies making it hard to find critical points in our library.  Further analyses are needed to check the output and implement filtering if necessary.  (see [issue](https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl/issues/214#issue-3574441926))

- [x] Tests fails:
    - [x] Julia nightly-ubuntu-latest: We use TestItemRunner.jl for structuring and running our testsets. Unfortunately, `Test.push_testset` has been removed in Julia nightly causing our tests to fail. We already reported the [issue in the TestItemRunner.jl repo](https://github.com/julia-vscode/TestItemRunner.jl/issues/113) and created a corresponding [issue](https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl/issues/218#issue-3586600920) in our repository. We hope this issue will be fixed in the near future.
    - [x] Julia lts-ubuntu-latest-x64 fails in our optimization tests.  The same test fails in some other Julia versions as well. Weirdly, the problem only occurs in our CI and it is hard to reproduce locally.  We noticed that there are ongoing changes in Optimization.jl package that might be related to the issue. We will investigate further and try to find a solution. See also [issue](https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl/issues/216#issue-3585882096).

- [x] BioMakie.jl:
    We are following the development of BioMakie.jl with great interest. We got the impression it has a slightly different focus (msa plot and alpha shapes). Our analysis require different functionalities. Consider the following scenario:
    You are visualizing an entire protein and then, you want to run the minimization algorithm only on the C-alpha atoms. The necessary filtering process are done with BiochemicalAlgorithms.jl and then you can visualize the movements of the atoms during the energy minimization steps.
    Consequently, the motivation for developing BiochemicalVisualization.jl is based on two aspects:
        1. We wanted to have a tool providing standard functionalities for protein bioinformatics  with a strong focus on representation used for modelling purposes. *ball-and-stick*, *van-der-waals*, *stick* representations are currently already supported and work on protein surfaces such as SAS is planned.
        2. We believe that a tight integration between the visualization tool and the underlying framework is required to ensure the necessary performance.

- [x] Mutable struct for system components:
     That is a great question, which is not so trivially answered.
     First, our data and analysis workflows do not fit with a bottom up approach: We typically read entire protein information (elements, positions) from raw data files and reconstruct the missing information like bonds or missing atoms (hydrogens). During our analysis, the data can change e.g., atom positions and velocities change.
     The `System` itself is the only container that is actually declared `mutable` and our other core structures are based on the `Tables.jl` interface. Simply put, the tables for atoms, bonds etc (like `AtomTable`, `BondTable`, etc.), act like DataFrames with type stability. Users are interacting with these mutable tables by immutable views on the tables (`Atom`, `Bond` etc.). To put it plainly, from user perspective it seems the components are mutable and for simplicity, in our documentation we describe them as mutable representations. From a technical perspective, the views are immutable and only the corresponding tables are mutable.
     Thank you for your suggestion to change to immutable structs. We could do this with the system with the two caveats: We would need to solve the indexing scheme and we would no longer be able to change the name of the system. I created an [issue](https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl/issues/215#issue-3574887706) for it and we will have a look at it.
