### 2025-11-06 #188 @kescobo
> Our analysis require different functionalities.

Re: BioMakie.jl - that all makes sense and I assumed as much, I just think a sentence or two in the manuscript along the lines of... "While there are existing julia packages (eg BioMakie.jl) to visualize protein structures, BiochemicalVisualization.jl does XYZ and its tight integration increases performace..." or something
