### 2025-11-12 b7e6aa6
[paper] added sentences about alternative protein viewers
