+++ paper/ref.bib
@@ -1,291 +1,299 @@
 @article{Kumar2021,
-	author = {Kumar, Nitin and Jain, Abhinav and Kumar, Ashwani and Kumar, Pankaj and Mishra, Anurag and Chaudhary, Kumardeep and Raghava, Gajendra P.S.},
-	title = {Structural bioinformatics resources for SARS-CoV-2},
-	journal = {Briefings in Bioinformatics},
-	volume = {22},
-	number = {3},
-	pages = {931-943},
-	year = {2021},
-	doi = {10.1093/bib/bbaa261},
-	url = {https://pmc.ncbi.nlm.nih.gov/articles/PMC7799268/},
-	pmid = {33141145},
-	pmc = {PMC7799268}
+  author  = {Kumar, Nitin and Jain, Abhinav and Kumar, Ashwani and Kumar, Pankaj and Mishra, Anurag and Chaudhary, Kumardeep and Raghava, Gajendra P.S.},
+  title   = {Structural bioinformatics resources for SARS-CoV-2},
+  journal = {Briefings in Bioinformatics},
+  volume  = {22},
+  number  = {3},
+  pages   = {931-943},
+  year    = {2021},
+  doi     = {10.1093/bib/bbaa261},
+  url     = {https://pmc.ncbi.nlm.nih.gov/articles/PMC7799268/},
+  pmid    = {33141145},
+  pmc     = {PMC7799268}
 }
 @article{Trott2010,
-	author = {Trott, Oliver and Olson, Arthur J.},
-	title = {AutoDock Vina: improving the speed and accuracy of docking with a new scoring function, efficient optimization, and multithreading},
-	journal = {Journal of Computational Chemistry},
-	volume = {31},
-	number = {2},
-	pages = {455-461},
-	year = {2010},
-	doi = {10.1002/jcc.21334}
+  author  = {Trott, Oliver and Olson, Arthur J.},
+  title   = {AutoDock Vina: improving the speed and accuracy of docking with a new scoring function, efficient optimization, and multithreading},
+  journal = {Journal of Computational Chemistry},
+  volume  = {31},
+  number  = {2},
+  pages   = {455-461},
+  year    = {2010},
+  doi     = {10.1002/jcc.21334}
 }
 
 @article{Morris2009AutoDock4AA,
-	title={AutoDock4 and AutoDockTools4: Automated docking with selective receptor flexibility},
-	author={Morris, Garrett M. and Huey, Ruth and Lindstrom, William and Sanner, Michel F. and Belew, Richard K. and Goodsell, David S. and Olson, Arthur J.},
-	journal={Journal of Computational Chemistry},
-	year={2009},
-	volume={30},
-	number={16},
-	pages={2785-2791},
-	doi={10.1002/jcc.21256}
+  title   = {AutoDock4 and AutoDockTools4: Automated docking with selective receptor flexibility},
+  author  = {Morris, Garrett M. and Huey, Ruth and Lindstrom, William and Sanner, Michel F. and Belew, Richard K. and Goodsell, David S. and Olson, Arthur J.},
+  journal = {Journal of Computational Chemistry},
+  year    = {2009},
+  volume  = {30},
+  number  = {16},
+  pages   = {2785-2791},
+  doi     = {10.1002/jcc.21256}
 }
 @article{berman2000protein,
-	title={The Protein Data Bank},
-	author={Berman, Helen M. and Westbrook, John and Feng, Zukang and Gilliland, Gary and Bhat, T. N. and Weissig, Helge and Shindyalov, Ilya N. and Bourne, Philip E.},
-	journal={Nucleic acids research},
-	volume={28},
-	number={1},
-	pages={235--242},
-	year={2000},
[…]
