### 2025-11-12 #188 @jeleclaire
> > Our analysis require different functionalities.
>
> Re: BioMakie.jl - that all makes sense and I assumed as much, I just think a sentence or two in the manuscript along the lines of... "While there are existing julia packages (eg BioMakie.jl) to visualize protein structures, BiochemicalVisualization.jl does XYZ and its tight integration increases performace..." or something

@kescobo
Thanks for pointing this out. I agree completely. This should be and is now in the manuscript.
I integrated the changes in the [application section about BiochemicalVisualization.jl](https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl/commit/b7e6aa673c90c8c33a8d21872014318716b8dbea).
