### 2025-11-18 380b485
CI: reenable Julia nightly testing
This reverts commit df358570a72a51fcc1f30f406e339a74c00f08dd.

Closes: #218
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
