### 2025-11-18 858cef1
Hotfix precompilation failures on Julia LTS (#220)
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
