### 2025-11-18 7dd8480
Hotfix precompilation failures on Julia LTS (#220)
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
