### 2025-11-18 b5137fd
CI: replace julia-actions/julia-downgrade-compat
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
