+++ .github/workflows/Downgrade.yml
@@ -22,7 +22,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - 'lts'
+          - '1.10'
           - '1.12'
         os:
           - ubuntu-latest
@@ -31,6 +31,8 @@ jobs:
         include:
           - version: '1.12'
             experimental: true
+    env:
+      RESOLVER_PATH: "/tmp/resolver"
     steps:
       - uses: actions/checkout@v5
       - uses: julia-actions/setup-julia@v2
@@ -38,9 +40,15 @@ jobs:
           version: ${{ matrix.version }}
           arch: ${{ matrix.arch }}
       - uses: julia-actions/cache@v2
-      - uses: julia-actions/julia-downgrade-compat@v1
-        with:
-          skip: LinearAlgebra,Printf,Statistics,Test
+      - name: Setup Resolver.jl
+        run: |
+          if [ ! -d "${RESOLVER_PATH}" ]; then
+            git clone https://github.com/StefanKarpinski/Resolver.jl.git "${RESOLVER_PATH}"
+            julia --project="${RESOLVER_PATH}/bin" -e 'using Pkg; Pkg.instantiate()'
+          fi
+      - name: Run resolver.jl
+        run: |
+          julia --project="${RESOLVER_PATH}/bin" ${RESOLVER_PATH}/bin/resolve.jl . --min-minor=@deps --julia=${{ matrix.version }}
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
         with:
