### 2025-11-18 3aae23b
CI: add additional Resolver.jl options
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
