+++ .github/workflows/Downgrade.yml
@@ -53,3 +53,5 @@ jobs:
       - uses: julia-actions/julia-runtest@v1
         with:
           coverage: false
+          allow_reresolve: false
+          force_latest_compatible_version: false
