### 2025-11-18 136b048
Test: exclude hotfixed IntervalSets from stale deps check
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
