+++ test/aqua.jl
@@ -1,5 +1,7 @@
 @testitem "Aqua" begin
     using Aqua
 
-    Aqua.test_all(BiochemicalAlgorithms)
+    # Ignore hotfixed package IntervalSets (https://github.com/hildebrandtlab/BiochemicalAlgorithms.jl/pull/221)
+    # for downgrade checks where the package is not pulled in (e.g. via MakieCore)
+    Aqua.test_all(BiochemicalAlgorithms; stale_deps = (; ignore = [:IntervalSets]))
 end
