### 2025-11-19 a421c8a
CompatHelper: bump compat for IntervalSets to 0.7, (keep existing compat)
