+++ Project.toml
@@ -51,7 +51,7 @@ Dates = "1.10"
 DocStringExtensions = "0.9"
 EnumX = "1"
 Graphs = "1.12"
-IntervalSets = "0.7 - 0.7.11"
+IntervalSets = "0.7 - 0.7.11, 0.7"
 JSON3 = "1.10"
 LinearAlgebra = "1.10"
 Mendeleev = "1"
