+++ .github/workflows/DCO.yml
@@ -5,7 +5,7 @@ jobs:
   dco-check:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v5
+    - uses: actions/checkout@v6
     - name: Set up Python 3.x
       uses: actions/setup-python@v6
       with:
