+++ .github/workflows/Downgrade.yml
@@ -34,7 +34,7 @@ jobs:
     env:
       RESOLVER_PATH: "/tmp/resolver"
     steps:
-      - uses: actions/checkout@v5
+      - uses: actions/checkout@v6
       - uses: julia-actions/setup-julia@v2
         with:
           version: ${{ matrix.version }}
