### 2025-11-24 0b0d4af
Remove hotfix for precompilation failures on Julia LTS
This reverts commit 7dd8480178d56ca302cc2cc710ec8270e86125fd.

Closes: #220
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
