+++ Project.toml
@@ -15,7 +15,6 @@ Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 EnumX = "4e289a0a-7415-4d19-859d-a7e5c4648b56"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-IntervalSets = "8197267c-284f-5f27-9208-e0e47529a953"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Mendeleev = "c116f080-063d-490a-9873-2b5b2cce4c34"
@@ -51,7 +50,6 @@ Dates = "1.10"
 DocStringExtensions = "0.9"
 EnumX = "1"
 Graphs = "1.12"
-IntervalSets = "0.7 - 0.7.11"
 JSON3 = "1.10"
 LinearAlgebra = "1.10"
 Mendeleev = "1"
