### 2025-11-24 724bce0
Remove hotfix for precompilation failures on Julia LTS
This reverts commit 858cef16f688c0cbc147bb324fd2a5aae98d6bd1.

Closes: #220
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
