+++ Project.toml
@@ -14,7 +14,6 @@ DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 EnumX = "4e289a0a-7415-4d19-859d-a7e5c4648b56"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-IntervalSets = "8197267c-284f-5f27-9208-e0e47529a953"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Mendeleev = "c116f080-063d-490a-9873-2b5b2cce4c34"
@@ -47,7 +46,6 @@ DataStructures = "0.18.13, 0.19"
 DocStringExtensions = "0.9"
 EnumX = "1"
 Graphs = "1.8"
-IntervalSets = "0.7 - 0.7.11"
 JSON3 = "1.9.1"
 LinearAlgebra = "1.10"
 Mendeleev = "1"
