### 2025-11-24 387da70
Revert "Test: exclude hotfixed IntervalSets from stale deps check"
This reverts commit 136b0480f0ef483ed0b736f12a288ace3cf8cec4.

Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
