### 2025-12-01 fbf8771
Separate gradient and loss function from optimize_structure
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
