### 2025-12-01 e54b726
added test case for validation of gradients with FiniteDifferences.jl
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
