+++ Project.toml
@@ -49,6 +49,7 @@ DataStructures = "0.18.13, 0.19"
 Dates = "1.10"
 DocStringExtensions = "0.9"
 EnumX = "1"
+FiniteDifferences = "0.12.33"
 Graphs = "1.12"
 JSON3 = "1.10"
 LinearAlgebra = "1.10"
@@ -77,8 +78,9 @@ julia = "1.10"
 
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
+FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TestItemRunner = "f8b46487-2199-4994-9208-9a1283c18c0a"
 
 [targets]
-test = ["Aqua", "Test", "TestItemRunner"]
+test = ["Aqua", "FiniteDifferences", "Test", "TestItemRunner"]
