### 2025-12-01 4353cb6
Implement bounding box computation
Signed-off-by: Andreas Hildebrandt <andreas.hildebrandt@uni-mainz.de>
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
