+++ docs/src/public/structureanalysis.md
@@ -1,0 +1,13 @@
+# Structure analysis
+```@meta
+CurrentModule = BiochemicalAlgorithms
+```
+
+```@index
+Pages = ["structureanalysis.md"]
+```
+
+## Bounding box computation
+```@docs
+compute_bounding_box
+```
