### 2025-12-01 703f592
Implement bounding box computation
Signed-off-by: Andreas Hildebrandt <andreas.hildebrandt@uni-mainz.de>
