+++ docs/src/public/structureanalysis.md
@@ -1,0 +1,14 @@
+# Preprocessing
+```@meta
+CurrentModule = BiochemicalAlgorithms
+```
+
+```@index
+Pages = ["structureanalysis.md"]
+```
+
+## Bounding box computation
+```@docs
+compute_bounding_box
+```
+
