### 2025-12-03 aed8ee1
CI: add Julia pre (1.13)
Signed-off-by: Thomas Kemmer <thomas@binaryden.de>
