### 2025-12-03 277cc57
draft for minibatching
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
