+++ test/optimization/test_optimize_structure.jl
@@ -25,3 +25,17 @@ end
     @test Int(optimize_hydrogen_positions!(ff).retcode) == 1
     @test compute_energy!(ff) ≈ 1421.212f0
 end
+
+@testitem "Optimize with minibatching" tags = [:skip_ci] begin
+    sys = load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
+
+    fdb = FragmentDB()
+    normalize_names!(sys, fdb)
+    reconstruct_fragments!(sys, fdb)
+    build_bonds!(sys, fdb)
+
+    ff = AmberFF(sys)
+    @test compute_energy!(ff) ≈ 1425.5979f0
+    @test Int(optimize_structure!(ff; minibatching=true).retcode) == 1
+    @test compute_energy!(ff) ≈ -374.3136f0
+end
