+++ Project.toml
@@ -27,6 +27,7 @@ PhysicalConstants = "5ad8b20f-a522-5ce9-bfc9-ddf1d5bda6ab"
 PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Quaternions = "94ee1d12-ae83-5a48-8b1c-48b8ff168ae0"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Rotations = "6038ab10-8711-5258-84ad-4b1120ba62dc"
 Scanf = "6ef1bc8b-493b-44e1-8d40-549aa65c4b41"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
@@ -62,6 +63,7 @@ PhysicalConstants = "0.2.3"
 PrettyTables = "3"
 Printf = "1.10"
 Quaternions = "0.7.6"
+Random = "1.11.0"
 Rotations = "1.3.4"
 Scanf = "0.5.4"
 StaticArrays = "1.9.8"
