### 2025-12-04 3523433
refactor code
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
