+++ src/forcefields/common/forcefield.jl
@@ -11,7 +11,7 @@ export
     count_warnings,
     print_warnings
 
-const force_prefactor = ustrip(u"kJ/mol/angstrom"/Constants.N_A |> u"N")
+const force_prefactor = ustrip(u"kJ/mol/angstrom" / Constants.N_A |> u"N")
 
 abstract type AbstractForceFieldComponent{T<:Real} end
 
@@ -19,10 +19,10 @@ abstract type AbstractForceFieldComponent{T<:Real} end
     name::String
     system::AbstractAtomContainer{T}
     parameters::AbstractForceFieldParameters
-    options::Dict{Symbol, Any}
-    atom_type_templates::Dict{String, AtomTypeTemplate{T}}
+    options::Dict{Symbol,Any}
+    atom_type_templates::Dict{String,AtomTypeTemplate{T}}
     components::Vector{AbstractForceFieldComponent{T}}
-    energy::Dict{String, T}
+    energy::Dict{String,T}
     unassigned_atoms::Vector{Atom{T}}
     constrained_atoms::Vector{Int}
     warnings::Vector{String}
@@ -31,38 +31,38 @@ abstract type AbstractForceFieldComponent{T<:Real} end
         name::String,
         system::AbstractAtomContainer{T},
         parameters::AbstractForceFieldParameters,
-        options::Dict{Symbol, Any} = Dict{Symbol, Any}(),
-        atom_type_templates::Dict{String, AtomTypeTemplate{T}} = Dict{String, AtomTypeTemplate{T}}(),
-        components::Vector{AbstractForceFieldComponent{T}} = Vector{AbstractForceFieldComponent{T}}(),
-        energy::Dict{String, T} = Dict{String, T}(),
-        unassigned_atoms::Vector{Atom{T}} = Vector{Atom{T}}(),
-        constrained_atoms::Vector{Int} = Vector{Int}(),
-        warnings::Vector{String} = Vector{String}()
+        options::Dict{Symbol,Any}=Dict{Symbol,Any}(),
+        atom_type_templates::Dict{String,AtomTypeTemplate{T}}=Dict{String,AtomTypeTemplate{T}}(),
+        components::Vector{AbstractForceFieldComponent{T}}=Vector{AbstractForceFieldComponent{T}}(),
+        energy::Dict{String,T}=Dict{String,T}(),
+        unassigned_atoms::Vector{Atom{T}}=Vector{Atom{T}}(),
+        constrained_atoms::Vector{Int}=Vector{Int}(),
+        warnings::Vector{String}=Vector{String}()
     ) where T
-        new{T}(name, system, parameters, options, atom_type_templates, components, energy, unassigned_atoms,constrained_atoms, warnings)
+        new{T}(name, system, parameters, options, atom_type_templates, components, energy, unassigned_atoms, constrained_atoms, warnings)
     end
 end
 
-function init_atom_types(params::AbstractForceFieldParameters, ::Type{T}=Float32) where {T <: Real}
+function init_atom_types(params::AbstractForceFieldParameters, ::Type{T}=Float32) where {T<:Real}
     tpl_section = extract_section(params, "ChargesAndTypeNames")
 
-    unit_q  = get(tpl_section.properties, "unit_q", "e_au")
+    unit_q = get(tpl_section.properties, "unit_q", "e_au")
 
     # UnitfulAtomic uses e_au for e0
     if unit_q == "e0"
         unit_q = "e_au"
     end
 
-    q_factor  = ustrip((1uparse(unit_q; unit_context=UnitfulAtomic))  |> u"e_au")
[…]
