### 2025-12-16 9330690
removed force_prefactor
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
