+++ test/forcefields/AMBER/test_amberff.jl
@@ -1,6 +1,10 @@
 @testitem "AmberFF" begin
+    using Unitful
+    const prefactor = ustrip(u"kJ/mol/angstrom"/Unitful.Na |> u"N")
+
+
     function _rms_F(sys::System{T}) where T
-        √(sum(squared_norm.(atoms(sys).F)) / (3 * natoms(sys))) / T(BiochemicalAlgorithms.force_prefactor)
+        √(sum(squared_norm.(atoms(sys).F)) / (3 * natoms(sys)))
     end
 
     let p = load_pdb(ball_data_path("../test/data/AlaAla.pdb"))
@@ -22,7 +26,7 @@
         @test a_ff.energy["Electrostatic"]    ≈ -85.1466827f0
 
         compute_forces!(a_ff)
-        @test _rms_F(p) ≈ 1703.33f0
+        @test _rms_F(p) ≈ 1682.0159f0
     end
 
     # BALL tests
@@ -52,7 +56,7 @@
             @test _rms_F(sys) ≈ T(228.72667607023757)
         end
 
-        # Energy test 2 (Bend) [AMBER91]
+     # Energy test 2 (Bend) [AMBER91]
         let sys = load_hinfile(ball_data_path("../test/data/AmberFF_test_2.hin"), T)
             normalize_names!(sys, fdb)
 
@@ -125,11 +129,11 @@
             @test ff.energy["Electrostatic"] ≈ T(-346.8087761352619)
 
             compute_forces!(ff)
-            @test _rms_F(sys) ≈ T(35.035943894791984)
+            @test _rms_F(sys) ≈ T(24.398090094191605)
         end
 
         # Energy test 5 (AlaGlySer) [AMBER96]
-        let sys = load_hinfile(ball_data_path("../test/data/AlaGlySer.hin"), T)
+          let sys = load_hinfile(ball_data_path("../test/data/AlaGlySer.hin"), T)
             normalize_names!(sys, fdb)
 
             ff = AmberFF(sys, ff_amber96; assign_charges = false, assign_typenames = false)
@@ -144,7 +148,7 @@
             @test ff.energy["Electrostatic"] ≈ T(57.20358862687402)
 
             compute_forces!(ff)
-            @test _rms_F(sys) ≈ T(40.63998952652902)
+            @test _rms_F(sys) ≈ T(28.500231091492697)
         end
 
         # Energy test 6 (AlaGlySer) [AMBER94]
@@ -163,7 +167,7 @@
             @test ff.energy["Electrostatic"] ≈ T(-163.88720300090736)
 
             compute_forces!(ff)
-            @test _rms_F(sys) ≈ T(41.758379137558485)
+            @test _rms_F(sys) ≈ T(29.38050165182444 )
         end
 
         # Force test 1 (Torsion) [AMBER94]
@@ -179,14 +183,14 @@
             @test only(ff.components) isa TorsionComponent{T}
 
             compute_forces!(ff)
-            @test atom_by_name(sys, "C").F ≈ T[5.65872e-10,7.01203e-13,-3.2631e-10] atol = 1e-12
-            @test atom_by_name(sys, "O").F ≈ T[-3.76732e-10,2.77556e-17,2.17502e-10] atol = 1e-12
-            @test atom_by_name(sys, "N").F ≈ T[9.69443e-11,2.64282e-10,9.07063e-11] atol = 1e-12
-            @test atom_by_name(sys, "H").F ≈ T[-2.86084e-10,-2.64983e-10,1.81024e-11] atol = 1e-12
+            @test atom_by_name(sys, "C").F ≈ T[34.07756609401984, 0.04222799449598291, -19.650846410165727] atol = 1e-4
[…]
