+++ tutorials/force_field_ops.qmd
@@ -57,6 +57,7 @@ The force vectors computed here are directly stored in the underlying system:
 atoms(sys).F   # or, equivalently, atoms(ff.system).F
 ```
 
+The force vectors are given in units of kJ/(mol·\AA).  In previous versions, forces were computed in Newton. 
 
 ## Potential energy computation
 
