### 2025-12-16 9569862
removed force_prefactor
Signed-off-by: Jenny Leclaire <leclaire@uni-mainz.de>
