+++ paper/ref.bib
@@ -121,7 +121,7 @@
   pages   = {4897-4909},
   year    = {2024},
   issn    = {2041-6520},
-  doi     = {https://doi.org/10.1039/d3sc05230c},
+  doi     = {10.1039/d3sc05230c},
   url     = {https://www.sciencedirect.com/science/article/pii/S2041652024003845}
 }
 
