### 2025-12-18 #188 @fcdimitr
Thank you all! @jeleclaire  Could you please check the Zenodo archive that is required for acceptance?
You need to follow [the instructions here](https://juliacon.github.io/proceedings-guide/author/#publication_phase) so that there is a permanent DOI for the final (accepted) version of the code & paper.
Then, we can accept it and close this issue. Thanks!
