### 2025-12-20 #188 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
