### 2025-12-20 #188 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
